% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agepro_options_flags.R
\name{agepro_options_flags}
\alias{agepro_options_flags}
\title{Class container that encapsulates logical flags to enable AGEPRO
user-defined options}
\description{
Encapsulates logical flags indicating that an optional AGEPRO's option can
be used: these options are percentile summary (PERC), scaling factors
(SCALE), biological reference points (REFPOINT), maximum bounds (BOUNDS), and
retrospective adjustment (RETROADJUST).
}
\details{
Associated with AGEPRO's output options (OPTIONS) are additional optional
options:

\describe{
\item{PERC}{user_percentile_summary}
\item{REFPOINT}{reference_points}
\item{SCALE}{scaling_factors}
\item{BOUNDS}{max_bounds}
\item{RETROADJUST}{retrospective_adjustment}
}

The AGEPRO input file format recognizes these optional keyword parameters.
At initialization, all option flags will be set to FALSE. To "enable" an
optional option or set it to TRUE, assign a value to the optional option's
field. For example, if the flag to enable for percentile summary
(\code{enable_user_percentile_summary}) is FALSE: set value \code{report_percentile},
Then it will be TRUE
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{enable_user_percentile_summary}}{\link[base:logical]{Logical} flag to allow percentile summary of the key
results in the output file.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-agepro_options_flags-new}{\code{agepro_options_flags$new()}}
\item \href{#method-agepro_options_flags-set_flag_user_percentile_summary}{\code{agepro_options_flags$set_flag_user_percentile_summary()}}
\item \href{#method-agepro_options_flags-clone}{\code{agepro_options_flags$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-agepro_options_flags-new"></a>}}
\if{latex}{\out{\hypertarget{method-agepro_options_flags-new}{}}}
\subsection{Method \code{new()}}{
Initialize the class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{agepro_options_flags$new(enable_percentile = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{enable_percentile}}{\link[base:logical]{Logical} flag to enable percentile summary of the key
results in the output file. FALSE by default.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-agepro_options_flags-set_flag_user_percentile_summary"></a>}}
\if{latex}{\out{\hypertarget{method-agepro_options_flags-set_flag_user_percentile_summary}{}}}
\subsection{Method \code{set_flag_user_percentile_summary()}}{
Toggles the logical flag to enable user_percentile summary
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{agepro_options_flags$set_flag_user_percentile_summary(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{Logical value}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-agepro_options_flags-clone"></a>}}
\if{latex}{\out{\hypertarget{method-agepro_options_flags-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{agepro_options_flags$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
