% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AgeproModel.R
\name{AgeproModel}
\alias{AgeproModel}
\title{R6 class representing AGEPRO model}
\description{
AGEPRO model contains the projection time horizon, age class range, number
of fleets, recruitment, and unertainties
}
\details{
AGEPRO performs stochastic projections on exploited fisheries stock to
determine age-structured population over a time period. Brodziak, 2022
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{yr_start}}{First Year in Projection}

\item{\code{yr_end}}{Last Year in Projection}

\item{\code{age_begin}}{First Age Class}

\item{\code{age_end}}{Last Age Class}

\item{\code{num_fleets}}{Number of Fleets}

\item{\code{num_rec_models}}{Number of Recruitment Models}

\item{\code{num_pop_sims}}{Number of Population Simulations}

\item{\code{discards}}{Are discards present?}

\item{\code{seed}}{Psuedorandom number seed}

\item{\code{inp_file}}{File name of opened/imported input file}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{num_years}}{Determines the number of years in projection by the (absolute) difference
of the last and first year of projection.}

\item{\code{num_ages}}{Determines number of ages by the (absolute) difference of the first and
last age class.}

\item{\code{seq_years}}{Returns a sequence of years from First year of projection}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AgeproModel-new}{\code{AgeproModel$new()}}
\item \href{#method-AgeproModel-get_json}{\code{AgeproModel$get_json()}}
\item \href{#method-AgeproModel-clone}{\code{AgeproModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AgeproModel-new"></a>}}
\if{latex}{\out{\hypertarget{method-AgeproModel-new}{}}}
\subsection{Method \code{new()}}{
Starts an instances of the AGEPRO Model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AgeproModel$new(
  yr_start = 0,
  yr_end = 1,
  age_begin = 0,
  age_end = 1,
  num_fleets = 1,
  num_rec_models = 1,
  num_pop_sims = 1000,
  discards = FALSE,
  seed = 0
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{yr_start}}{First Year of Projection}

\item{\code{yr_end}}{Last Year of Projection}

\item{\code{age_begin}}{age begin}

\item{\code{age_end}}{age end}

\item{\code{num_fleets}}{Number of fleets}

\item{\code{num_rec_models}}{Number of Recruit Modles}

\item{\code{num_pop_sims}}{Number of population sims}

\item{\code{discards}}{discards}

\item{\code{seed}}{Random Number seed}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AgeproModel-get_json"></a>}}
\if{latex}{\out{\hypertarget{method-AgeproModel-get_json}{}}}
\subsection{Method \code{get_json()}}{
Get json
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AgeproModel$get_json()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AgeproModel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AgeproModel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AgeproModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
