% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ageproR-package.R
\docType{package}
\name{ageproR-package}
\alias{ageproR}
\alias{ageproR-package}
\title{ageproR: R package interface for AGEPRO (Age Structured Projection Model) data}
\description{
Object-oriented data structures based on the input file format defined in the Age Structured Projection Model Reference Manual. Includes support modules for AGEPRO input file support, and experimental JSON input file format.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nmfs-ost/ageproR}
  \item Report bugs at \url{https://github.com/nmfs-ost/ageproR/issues}
}

}
\author{
\strong{Maintainer}: Eric Fletcher \email{eric.fletcher@noaa.gov}

Authors:
\itemize{
  \item Eric Fletcher \email{eric.fletcher@noaa.gov}
  \item Jon Brodziak
}

}
\keyword{internal}
