% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assert_numeric_substrings}
\alias{assert_numeric_substrings}
\title{Asserts if all substrings of AGEPRO's input data file line can be numeric.}
\usage{
assert_numeric_substrings(inp_line)
}
\arguments{
\item{inp_line}{Line read from the AGEPRO Input File}
}
\value{
Converts the input data line string vector as numeric.
}
\description{
Validates the string vector via `grepl` if all values match the digit
character class. Function will throw an exception if non digit characters
were found.
}
\keyword{internal}
