% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{capture_output_as_message}
\alias{capture_output_as_message}
\title{converts output as messages}
\usage{
capture_output_as_message(x)
}
\arguments{
\item{x}{Vector output object}
}
\description{
Helper function to format vector output, such as data.frames, matrices. or
lists to message format.
}
