% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model.R
\name{check_model_bootstrap}
\alias{check_model_bootstrap}
\title{Helper function to check bootstrap file locations prior AGEPRO calcuation engine launch.}
\usage{
check_model_bootstrap(model)
}
\arguments{
\item{model}{AGEPRO model class for AGEPRO Input File format (\code{agepro_inp_model})}
}
\description{
Helper function to check bootstrap file locations prior AGEPRO calcuation engine launch.
}
