% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{check_model_num_vector_args}
\alias{check_model_num_vector_args}
\alias{assert_model_num_vector_args}
\title{Recruitment Model Number Parameter validation}
\usage{
check_model_num_vector_args(x)

assert_model_num_vector_args(x, .var.name = checkmate::vname(x), add = NULL)
}
\arguments{
\item{x}{object to check}

\item{.var.name}{\code{character(1)}\cr
Name of the checked object to print in assertions. Defaults to
the heuristic implemented in
\code{\link[checkmate:vname]{checkmate::vname}}.}

\item{add}{\code{checkmate::AssertCollection}\cr
Collection to store assertion messages. See
\code{\link[checkmate:AssertCollection]{checkmate::AssertCollection}}}
}
\description{
Custom validation to check
\href{../../ageproR/html/agepro_model.html#method-set_recruit_model}{\code{agepro_model$set_recruit_model()}}
arguments to see if multiple recruit numbers is passed as a single vector
or seen as a list of multiple arguments.

If the input value is passed as a list of multiple arguments,
this function will "throw" a message of the issue and possible resolution.

\link{assert_model_num_vector_args} wraps
\link{check_model_num_vector_args} as a custom checkmate assertion via \link[checkmate:makeAssertion]{checkmate::makeAssertion}
}
