% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{check_model_num_vector_count}
\alias{check_model_num_vector_count}
\alias{assert_model_num_vector_count}
\title{Recruitment model number vector count validation}
\usage{
check_model_num_vector_count(x, num_recruit_models)

assert_model_num_vector_count(
  x,
  num_recruit_models,
  .var.name = checkmate::vname(x),
  add = NULL
)
}
\arguments{
\item{x}{Object to check}

\item{num_recruit_models}{Number of recruitment models AGEPRO model was intialized with#'}

\item{.var.name}{\code{character(1)}\cr
Name of the checked object to print in assertions. Defaults to
the heuristic implemented in
\code{\link[checkmate:vname]{checkmate::vname}}.}

\item{add}{\code{checkmate::AssertCollection}\cr
Collection to store assertion messages. See
\code{\link[checkmate:AssertCollection]{checkmate::AssertCollection}}}
}
\description{
Checks if input model number matches the number of recruitment models of
the model.
}
