% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{check_perc_active_binding}
\alias{check_perc_active_binding}
\alias{assert_perc_active_binding}
\title{Validation for PERC active binding in agepro_model}
\usage{
check_perc_active_binding(x)

assert_perc_active_binding(x, .var.name = checkmate::vname(x), add = NULL)
}
\arguments{
\item{x}{Object to Check}

\item{.var.name}{\code{character(1)}\cr
Name of the checked object to print in assertions. Defaults to
the heuristic implemented in
\code{\link[checkmate:vname]{checkmate::vname}}.}

\item{add}{\code{checkmate::AssertCollection}\cr
Collection to store assertion messages. See
\code{\link[checkmate:AssertCollection]{checkmate::AssertCollection}}}
}
\description{
Custom validation procedure to check if input value matches the structure
of the user_percentile_summary R6Class. It will also catch single numeric
input values presuming that the active binder sets the report_percentile.
}
