% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_proj_years_parameter}
\alias{check_proj_years_parameter}
\title{Validates the usage of the 'projection years' parameter.}
\usage{
check_proj_years_parameter(proj_years)
}
\arguments{
\item{proj_years}{Projection year parameter. May be a numeric vector or a
\code{\link[=projection_years]{projection_years}}}
}
\description{
#' If \code{proj_years} parameter is a
\link[=projection_years]{projection_years class}, then it will return
that value. Otherwise, it will create a new \code{projection_years} class based
on the param value passed.
}
