% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{check_proj_years_sequence}
\alias{check_proj_years_sequence}
\alias{assert_proj_years_sequence}
\title{Projection year sequence vector validation}
\usage{
check_proj_years_sequence(x)

assert_proj_years_sequence(x, .var.name = checkmate::vname(x), add = NULL)
}
\arguments{
\item{x}{Object to check}

\item{.var.name}{\code{character(1)}\cr
Name of the checked object to print in assertions. Defaults to
the heuristic implemented in
\code{\link[checkmate:vname]{checkmate::vname}}.}

\item{add}{\code{checkmate::AssertCollection}\cr
Collection to store assertion messages. See
\code{\link[checkmate:AssertCollection]{checkmate::AssertCollection}}}
}
\description{
Custom validation to check projection_years sequence vector is
properly incremented by 1, Can be Wrapped as a custom checkmate assertion.
}
