% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cli_keyword_heading}
\alias{cli_keyword_heading}
\title{Command-line interface header for AGEPRO Keyword parameters}
\usage{
cli_keyword_heading(keyword, heading_color = "cyan")
}
\arguments{
\item{keyword}{Text of the header}

\item{header_color}{Text Color, R color, or HTML hexidecimal color.}
}
\description{
Creates an custom header with double-lines (colored in cyan by default), via
cli library.
}
\keyword{internal}
