% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruit_model.R
\name{deprecated_recruit_model_9}
\alias{deprecated_recruit_model_9}
\title{DEPRECATED Recruitment Model #9
(Time-Varying Empirical Recruitment Distribution)}
\description{
Handles an instance for deprecated recruitment model #9.
}
\section{Super class}{
\code{\link[ageproR:recruit_model]{recruit_model}} -> \code{deprecated_recruit_model_9}
}
\section{Active bindings}{
  \if{html}{\out{<div class="r6-active-bindings">}}
  \describe{
    \item{\code{json_recruit_data}}{Function container to export recruitment model data to experimental
jSon input model file. Because recruitment model #9 is DEPRECATED, and
not used in the AGEPRO calculation engine, an error will thrown.}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-deprecated_recruit_model_9-initialize}{\code{deprecated_recruit_model_9$new()}}
    \item \href{#method-deprecated_recruit_model_9-print}{\code{deprecated_recruit_model_9$print()}}
    \item \href{#method-deprecated_recruit_model_9-inp_lines_recruit_data}{\code{deprecated_recruit_model_9$inp_lines_recruit_data()}}
    \item \href{#method-deprecated_recruit_model_9-clone}{\code{deprecated_recruit_model_9$clone()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-deprecated_recruit_model_9-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-deprecated_recruit_model_9-initialize}{}}}
\subsection{\code{deprecated_recruit_model_9$new()}}{
  Initializes the class
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{deprecated_recruit_model_9$new()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-deprecated_recruit_model_9-print"></a>}}
\if{latex}{\out{\hypertarget{method-deprecated_recruit_model_9-print}{}}}
\subsection{\code{deprecated_recruit_model_9$print()}}{
  Prints out the error
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{deprecated_recruit_model_9$print(...)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{...}}{further arguments passed to or from other methods}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-deprecated_recruit_model_9-inp_lines_recruit_data"></a>}}
\if{latex}{\out{\hypertarget{method-deprecated_recruit_model_9-inp_lines_recruit_data}{}}}
\subsection{\code{deprecated_recruit_model_9$inp_lines_recruit_data()}}{
  Function container to export recruitment model data to AGEPRO input model
file lines, but since this model is DEPRECATED; not a valid recruitment
model type for the AGEPRO calculation engine, an error will thrown.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{deprecated_recruit_model_9$inp_lines_recruit_data(delimiter = " ")}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{delimiter}}{Character string delimiter separating values of AGEPRO input file line.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-deprecated_recruit_model_9-clone"></a>}}
\if{latex}{\out{\hypertarget{method-deprecated_recruit_model_9-clone}{}}}
\subsection{\code{deprecated_recruit_model_9$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{deprecated_recruit_model_9$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
