% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_error.R
\name{discard_fraction}
\alias{discard_fraction}
\title{Discard fraction of numbers at age}
\description{
Class Structure for discard faction at age AGEPRO Keyword parameter.
AGEPRO model must indicate \emph{discards are present}, enabled via general
options \link[=general_params]{discards_present field}.
}
\section{Super class}{
\code{\link[ageproR:process_error]{process_error}} -> \code{discard_fraction}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-discard_fraction-initialize}{\code{discard_fraction$new()}}
    \item \href{#method-discard_fraction-clone}{\code{discard_fraction$clone()}}
  }
}
\if{html}{\out{<details><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="ageproR" data-topic="process_error" data-id="create_parameter_table"><a href='../../ageproR/html/process_error.html#method-process_error-create_parameter_table'><code>process_error$create_parameter_table()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="ageproR" data-topic="process_error" data-id="get_inp_lines"><a href='../../ageproR/html/process_error.html#method-process_error-get_inp_lines'><code>process_error$get_inp_lines()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="ageproR" data-topic="process_error" data-id="print"><a href='../../ageproR/html/process_error.html#method-process_error-print'><code>process_error$print()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="ageproR" data-topic="process_error" data-id="read_inp_lines"><a href='../../ageproR/html/process_error.html#method-process_error-read_inp_lines'><code>process_error$read_inp_lines()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="ageproR" data-topic="process_error" data-id="read_inp_lines_cv_table"><a href='../../ageproR/html/process_error.html#method-process_error-read_inp_lines_cv_table'><code>process_error$read_inp_lines_cv_table()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="ageproR" data-topic="process_error" data-id="read_inp_lines_parameter_tables"><a href='../../ageproR/html/process_error.html#method-process_error-read_inp_lines_parameter_tables'><code>process_error$read_inp_lines_parameter_tables()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-discard_fraction-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-discard_fraction-initialize}{}}}
\subsection{\code{discard_fraction$new()}}{
  Initializes Class
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{discard_fraction$new(
  proj_years,
  num_ages,
  num_fleets,
  input_option = 0,
  time_varying = TRUE,
  enable_cat_print = TRUE
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{proj_years}}{\link[=projection_years]{Projection years}:
Input can be Sequence of years in from first to last year of
projection or the number of years in the time projection.}
      \item{\code{num_ages}}{Number of Age classes}
      \item{\code{num_fleets}}{Number of Fleets.}
      \item{\code{input_option}}{Option to indicate this parameter will be:
\itemize{
\item \code{0} By default, done interactively via interface.
\item \code{1} Imported via location of an existing data file.
}}
      \item{\code{time_varying}}{Logical flag that enables the stochastic parameter
to use as a time-varying array if TRUE (or 1). Otherwise, FALSE the
vector will cover "all years" of the projection. Default is TRUE.}
      \item{\code{enable_cat_print}}{Logical flag to show target function's \strong{cli} \code{\link[cli:cat_print]{cat_print}}
messages to be seen on console.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-discard_fraction-clone"></a>}}
\if{latex}{\out{\hypertarget{method-discard_fraction-clone}{}}}
\subsection{\code{discard_fraction$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{discard_fraction$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
