% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{div_keyword_header}
\alias{div_keyword_header}
\title{Command-line interface header for AGEPRO Keyword parameters}
\usage{
div_keyword_header(keyword, header_color = "cyan")
}
\arguments{
\item{keyword}{Text of the header}

\item{header_color}{Text Color, R color, or HTML hexidecimal color.}
}
\description{
Creates an custom header with double-lines (colored in cyan by default), via
cli library.
}
\keyword{internal}
