% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruit_model.R
\name{EmpiricalCDFModel}
\alias{EmpiricalCDFModel}
\title{Empirical CDF of Recruitment (Model #14)}
\description{
Empirical CDF of Recruitment (Model #14)

Empirical CDF of Recruitment (Model #14)
}
\section{Super classes}{
\code{\link[AGEPROjson:recruit_model]{AGEPROjson::recruit_model}} -> \code{\link[AGEPROjson:EmpiricalRecruitModel]{AGEPROjson::EmpiricalRecruitModel}} -> \code{EmpiricalCDFModel}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-EmpiricalCDFModel-new}{\code{EmpiricalCDFModel$new()}}
\item \href{#method-EmpiricalCDFModel-clone}{\code{EmpiricalCDFModel$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="AGEPROjson" data-topic="EmpiricalRecruitModel" data-id="new_obs_table"><a href='../../AGEPROjson/html/EmpiricalRecruitModel.html#method-EmpiricalRecruitModel-new_obs_table'><code>AGEPROjson::EmpiricalRecruitModel$new_obs_table()</code></a></span></li>
<li><span class="pkg-link" data-pkg="AGEPROjson" data-topic="EmpiricalRecruitModel" data-id="print"><a href='../../AGEPROjson/html/EmpiricalRecruitModel.html#method-EmpiricalRecruitModel-print'><code>AGEPROjson::EmpiricalRecruitModel$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="AGEPROjson" data-topic="EmpiricalRecruitModel" data-id="print_json"><a href='../../AGEPROjson/html/EmpiricalRecruitModel.html#method-EmpiricalRecruitModel-print_json'><code>AGEPROjson::EmpiricalRecruitModel$print_json()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EmpiricalCDFModel-new"></a>}}
\if{latex}{\out{\hypertarget{method-EmpiricalCDFModel-new}{}}}
\subsection{Method \code{new()}}{
Initialize the Empirical CDF Model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EmpiricalCDFModel$new(seq_years)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{seq_years}}{\cr
Array representing the Projection Time Horizon}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EmpiricalCDFModel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-EmpiricalCDFModel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EmpiricalCDFModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
