% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecruitModel.R
\name{EmpiricalDistributionModel}
\alias{EmpiricalDistributionModel}
\title{Empirical Recruitment Distribution (Model #3)}
\description{
Empirical Recruitment Distribution (Model #3)

Empirical Recruitment Distribution (Model #3)
}
\section{Super classes}{
\code{\link[AGEPROjson:RecruitModel]{AGEPROjson::RecruitModel}} -> \code{\link[AGEPROjson:EmpiricalRecruitModel]{AGEPROjson::EmpiricalRecruitModel}} -> \code{EmpiricalDistributionModel}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-EmpiricalDistributionModel-new}{\code{EmpiricalDistributionModel$new()}}
\item \href{#method-EmpiricalDistributionModel-clone}{\code{EmpiricalDistributionModel$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="AGEPROjson" data-topic="EmpiricalRecruitModel" data-id="new_obs_table"><a href='../../AGEPROjson/html/EmpiricalRecruitModel.html#method-EmpiricalRecruitModel-new_obs_table'><code>AGEPROjson::EmpiricalRecruitModel$new_obs_table()</code></a></span></li>
<li><span class="pkg-link" data-pkg="AGEPROjson" data-topic="EmpiricalRecruitModel" data-id="print_json"><a href='../../AGEPROjson/html/EmpiricalRecruitModel.html#method-EmpiricalRecruitModel-print_json'><code>AGEPROjson::EmpiricalRecruitModel$print_json()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EmpiricalDistributionModel-new"></a>}}
\if{latex}{\out{\hypertarget{method-EmpiricalDistributionModel-new}{}}}
\subsection{Method \code{new()}}{
Initialize the Empirical Recruitment Distribution Model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EmpiricalDistributionModel$new(seq_years)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{seq_years}}{\cr
Array representing the Projection Time Horizon}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EmpiricalDistributionModel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-EmpiricalDistributionModel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EmpiricalDistributionModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
