% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralParams.R
\name{GeneralParams}
\alias{GeneralParams}
\title{Input general model parameters}
\description{
Stores overall AGERPRO model project parameters including
the projection time horizon, the target model's age class range, the
quantity of recruitment models used, and fishery processes that affect
projections.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{yr_start}}{First Year in Projection}

\item{\code{yr_end}}{Last Year in Projection}

\item{\code{age_begin}}{First Age Class}

\item{\code{age_end}}{Last Age Class}

\item{\code{num_fleets}}{Number of Fleets}

\item{\code{num_rec_models}}{Number of Recruitment Models}

\item{\code{num_pop_sims}}{Number of Population Simulations}

\item{\code{discards}}{Are discards present?}

\item{\code{seed}}{Psuedorandom number seed}

\item{\code{inp_file}}{File name of opened/imported input file}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{num_years}}{Determines the number of years in projection by the (absolute) difference
of the last and first year of projection.}

\item{\code{num_ages}}{Determines number of ages by the (absolute) difference of the first and
last age class.}

\item{\code{seq_years}}{Returns a sequence of years from First year of projection}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GeneralParams-new}{\code{GeneralParams$new()}}
\item \href{#method-GeneralParams-clone}{\code{GeneralParams$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GeneralParams-new"></a>}}
\if{latex}{\out{\hypertarget{method-GeneralParams-new}{}}}
\subsection{Method \code{new()}}{
Starts an instances of the AGEPRO Model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeneralParams$new(
  yr_start = 0,
  yr_end = 1,
  age_begin = 0,
  age_end = 1,
  num_fleets = 1,
  num_rec_models = 1,
  num_pop_sims = 1000,
  discards = FALSE,
  seed = 0
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{yr_start}}{First Year of Projection}

\item{\code{yr_end}}{Last Year of Projection}

\item{\code{age_begin}}{age begin}

\item{\code{age_end}}{age end}

\item{\code{num_fleets}}{Number of fleets}

\item{\code{num_rec_models}}{Number of Recruit Modles}

\item{\code{num_pop_sims}}{Number of population sims}

\item{\code{discards}}{discards}

\item{\code{seed}}{Random Number seed}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GeneralParams-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GeneralParams-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeneralParams$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
