% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{get_r_frontend}
\alias{get_r_frontend}
\title{Retruns GUI Frontend of current R process.}
\usage{
get_r_frontend()
}
\description{
Helper function to get the GUI frontend attached to the current Rconsole process.
This informtion will be used to determine the rstudioapi, windows native base R, or
muiltiplatform tcl/tk.
}
