% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_file.R
\name{input_file}
\alias{input_file}
\title{AGEPRO Input File}
\description{
AGEPRO Input File

AGEPRO Input File
}
\details{
File Functionality is based on AGEPRO-CoreLib implementation
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-input_file-new}{\code{input_file$new()}}
\item \href{#method-input_file-read_inpfile}{\code{input_file$read_inpfile()}}
\item \href{#method-input_file-check_inpfile_version}{\code{input_file$check_inpfile_version()}}
\item \href{#method-input_file-read_inpfile_values}{\code{input_file$read_inpfile_values()}}
\item \href{#method-input_file-clone}{\code{input_file$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-input_file-new"></a>}}
\if{latex}{\out{\hypertarget{method-input_file-new}{}}}
\subsection{Method \code{new()}}{
Initializes the input file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{input_file$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-input_file-read_inpfile"></a>}}
\if{latex}{\out{\hypertarget{method-input_file-read_inpfile}{}}}
\subsection{Method \code{read_inpfile()}}{
Reads in AGEPRPO input file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{input_file$read_inpfile(inpfile)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{inpfile}}{input file}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-input_file-check_inpfile_version"></a>}}
\if{latex}{\out{\hypertarget{method-input_file-check_inpfile_version}{}}}
\subsection{Method \code{check_inpfile_version()}}{
Check Input File Version
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{input_file$check_inpfile_version(inpline)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{inpline}}{Input File Line}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-input_file-read_inpfile_values"></a>}}
\if{latex}{\out{\hypertarget{method-input_file-read_inpfile_values}{}}}
\subsection{Method \code{read_inpfile_values()}}{
Read Input file Values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{input_file$read_inpfile_values(inp_con)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{inp_con}}{File connection}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-input_file-clone"></a>}}
\if{latex}{\out{\hypertarget{method-input_file-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{input_file$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
