% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{invalid_path_message}
\alias{invalid_path_message}
\title{Invalid Path Message}
\usage{
invalid_path_message(x)
}
\arguments{
\item{x}{Filepath string}
}
\description{
Returns a reusable invalid Path Message
}
