% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{is_rstudio_desktop}
\alias{is_rstudio_desktop}
\title{Checks Rstudioapi if Rstudio Desktop is used.}
\usage{
is_rstudio_desktop()
}
\description{
vscode uses/emulates rstudioapi but not all features rstudioapi are
implemented. vscode's rstudio version information is set to '0'.
For Rstudio specific code, check for mode "desktop", and version > '0'
}
