% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model.R
\name{model_bootstrap_check}
\alias{model_bootstrap_check}
\title{Helper function to check bootstrap file locations prior AGEPRO calcuation engine launch.}
\usage{
model_bootstrap_check(model, bsn)
}
\arguments{
\item{model}{AGEPRO model class for AGEPRO Input File format (\code{agepro_inp_model})}

\item{bsn}{Bootstrap filepath}
}
\description{
Helper function to check bootstrap file locations prior AGEPRO calcuation engine launch.
}
