% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecruitModel.R
\name{NullRecruitModel}
\alias{NullRecruitModel}
\title{Null Recruitment UI Fallback Default}
\description{
Recruitment Model Description
}
\section{Super class}{
\code{\link[AGEPROjson:RecruitModel]{AGEPROjson::RecruitModel}} -> \code{NullRecruitModel}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-NullRecruitModel-new}{\code{NullRecruitModel$new()}}
\item \href{#method-NullRecruitModel-print}{\code{NullRecruitModel$print()}}
\item \href{#method-NullRecruitModel-clone}{\code{NullRecruitModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NullRecruitModel-new"></a>}}
\if{latex}{\out{\hypertarget{method-NullRecruitModel-new}{}}}
\subsection{Method \code{new()}}{
Initialize
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NullRecruitModel$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NullRecruitModel-print"></a>}}
\if{latex}{\out{\hypertarget{method-NullRecruitModel-print}{}}}
\subsection{Method \code{print()}}{
Prints out NULL Recruiment Model Data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NullRecruitModel$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{further arguments passed to or from other methods}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NullRecruitModel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-NullRecruitModel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NullRecruitModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
