% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{open_file}
\alias{open_file}
\title{Open file dialog or interface to interactively return file path.}
\usage{
open_file()
}
\arguments{
\item{inpfile}{input file name}
}
\description{
If Rconsole is currently running in Rstudio, it will use the rstudioapi
to show the file dialog window over the IDE.
}
\keyword{internal}
