% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{output_fn}
\alias{output_fn}
\title{Output filenames}
\usage{
output_fn()
}
\description{
Constructs output file name paths
}
\details{
Check if current working directory is a R project, in a R package source
directory, or has git root subdirectory. Using rprojroot find_root_file
criteria, a match will return a path if criteria matches. Otherwise,
tryCatch will catch the "not found" error and return FALSE. These values
will determine if the current working directory matches these three
condtions.
}
