% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{output_subdir}
\alias{output_subdir}
\title{Output Subdirectory}
\usage{
output_subdir(ignore_outdir = TRUE)
}
\arguments{
\item{ignore_outdir}{Append the output subdirectory to .gitignore and
.Rbuildignore. In case the target working directory only includes a
R project file and doesn't include git and R package files, this option is
not needed.}
}
\description{
If the current working directory has an instance or instances of either a
R project file, a git root subdirectory, or detected R package source files,
then the output will be pointed to the output subdirectory. The output
subdirectory will be created. In addition to this, it will add the output
subdirectory to \code{.gitignore} and \code{.Rbuildignore} if the git root
subdirectory and/or R package criteria was
found.
}
