% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print_parameter_table}
\alias{print_parameter_table}
\title{Print AGEPRO table-like values to the console}
\usage{
print_parameter_table(tbl, omit_rows = FALSE)
}
\arguments{
\item{tbl}{AGEPRO Keyword Parameter matrix/vector variable}

\item{omit_rows}{Logical flag, if \code{TRUE}, will print the first six rows,
via \code{\link[utils:head]{head()}}, to R console. In addition, the total number
of rows and rows omitted will be displayed. If \code{FALSE}, by default,
the matrix or vector prints normally.}
}
\description{
Helper function to print out AGEPRO keyword parameter's table-like matrix or
vector variables to console. Includes an option to show the first few rows
to the console.
}
