% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prompt_yes_no}
\alias{prompt_yes_no}
\title{Yes/No console Prompt}
\usage{
prompt_yes_no(question, default = NULL)
}
\arguments{
\item{question}{A character string used to prompt the user for input.}

\item{default}{Logical Value if user inputs prompt without typing anying.
Values can be TRUE (Yes), FALSE (No), or NULL. Default is NULL.}
}
\value{
Returns a Logical Value based ong prompt, if
\itemize{
\item \code{TRUE} User entered \code{y} (or \code{yes})
\item \code{FALSE} User entered \code{n} (or \code{no})
\item \code{NA} User entered \code{c} (or \code{cancel}, \code{q}, \code{quit})
}
}
\description{
Helper function to prompt a Yes/No prompt on the Rconsole
}
