% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{read_inp_numeric_line}
\alias{read_inp_numeric_line}
\title{Reads a line of numeric strings from the AGEPRO input file connection.}
\usage{
read_inp_numeric_line(inp_con)
}
\arguments{
\item{inp_con}{Open file connection to AGEPRO Input File.}
}
\description{
Reads in a line from the open file connection, splits the string
into substrings by whitespace, validates for numerical strings, and
then converts to numerical vector.
}
\keyword{internal}
