% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruit_model.R
\name{recruit_model}
\alias{recruit_model}
\title{Recruit Model}
\description{
Recruitment Model Description
}
\section{Active bindings}{
  \if{html}{\out{<div class="r6-active-bindings">}}
  \describe{
    \item{\code{model_num}}{Model number}

    \item{\code{model_group}}{Group type of Recruitment Model}

    \item{\code{model_name}}{Name of Recruitment Model}

    \item{\code{projected_years}}{Time Series of Projected Years}

    \item{\code{length_projected_years}}{Length of projected_years counted as the
the number of recruitment observations for some models.}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-recruit_model-initialize}{\code{recruit_model$new()}}
    \item \href{#method-recruit_model-clone}{\code{recruit_model$clone()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-recruit_model-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-recruit_model-initialize}{}}}
\subsection{\code{recruit_model$new()}}{
  Creates a new instance of this class
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{recruit_model$new()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-recruit_model-clone"></a>}}
\if{latex}{\out{\hypertarget{method-recruit_model-clone}{}}}
\subsection{\code{recruit_model$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{recruit_model$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
