% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecruitModel.R
\name{RecruitModel}
\alias{RecruitModel}
\title{Recruit Model}
\description{
Recruitment Model Description
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{model_num}}{Model number}

\item{\code{model_group}}{Group type of Recruitment Model}

\item{\code{model_name}}{Name of Recruitment Model}

\item{\code{seq_yrs}}{Time Series of Projected Years}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RecruitModel-new}{\code{RecruitModel$new()}}
\item \href{#method-RecruitModel-clone}{\code{RecruitModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RecruitModel-new"></a>}}
\if{latex}{\out{\hypertarget{method-RecruitModel-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RecruitModel$new(model_num, model_group)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model_num}}{\cr
AGEPRO Recruitment Model Number}

\item{\code{model_group}}{Model Type}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RecruitModel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RecruitModel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RecruitModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
