% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecruitModel.R
\name{RickerCurveModel}
\alias{RickerCurveModel}
\title{Ricker Curve #/ Lognormal Error (Model #6)}
\description{
Ricker Curve #/ Lognormal Error (Model #6)

Ricker Curve #/ Lognormal Error (Model #6)
}
\section{Super classes}{
\code{\link[AGEPROjson:RecruitModel]{AGEPROjson::RecruitModel}} -> \code{AGEPROjson::ParametricRecruitModel} -> \code{RickerCurveModel}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RickerCurveModel-new}{\code{RickerCurveModel$new()}}
\item \href{#method-RickerCurveModel-clone}{\code{RickerCurveModel$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="AGEPROjson" data-topic="ParametricRecruitModel" data-id="print"><a href='../../AGEPROjson/html/ParametricRecruitModel.html#method-ParametricRecruitModel-print'><code>AGEPROjson::ParametricRecruitModel$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="AGEPROjson" data-topic="ParametricRecruitModel" data-id="set"><a href='../../AGEPROjson/html/ParametricRecruitModel.html#method-ParametricRecruitModel-set'><code>AGEPROjson::ParametricRecruitModel$set()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RickerCurveModel-new"></a>}}
\if{latex}{\out{\hypertarget{method-RickerCurveModel-new}{}}}
\subsection{Method \code{new()}}{
Initalizes the Ricker Curve Model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RickerCurveModel$new(alpha = 0, beta = 0, variance = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{alpha}}{\cr
Stock Recruitment Parameter, alpha}

\item{\code{beta}}{\cr
Stock Recruitment Parameter, beta}

\item{\code{variance}}{\cr
variance}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RickerCurveModel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RickerCurveModel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RickerCurveModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
