% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{run_tcltk_dialog}
\alias{run_tcltk_dialog}
\title{Tcltk File Dialog helpers}
\usage{
run_tcltk_dialog(type = c("open", "save"), ext = c("All Files", ".*"))
}
\arguments{
\item{type}{Character option to request a "Open" or "Save file Dialog}

\item{ext}{Two item vector describing the filename description and filename extension respecively. By default it is assigned to \code{c("All Files", "*")}}
}
\value{
A Character vector representing the filepath of the chosen path from the file dialog.
}
\description{
Helper Function that allows clean Tcltk execution and teartown
}
\keyword{internal}
