% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dialog.R
\name{save_file_dialog}
\alias{save_file_dialog}
\title{Save file dialog or interface to interactively return file path.}
\usage{
save_file_dialog(filetype)
}
\arguments{
\item{filetype}{filename extension}
}
\description{
If Rconsole is currently running in Rstudio, it will use the rstudioapi
to show the file dialog window over the IDE.
}
\keyword{internal}
