% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_factors.R
\name{scale_factors}
\alias{scale_factors}
\title{Scale Factors}
\description{
Input information on scaling factors for biomass, recruitment, and stock size
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{keyword_name}}{AGEPRO keyword parameter name}

\item{\code{inp_keyword}}{Returns AGEPRO input-file formatted Parameter}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-scale_factors-clone}{\code{scale_factors$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-scale_factors-clone"></a>}}
\if{latex}{\out{\hypertarget{method-scale_factors-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{scale_factors$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
