% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_input.R
\name{set_bootstrap}
\alias{set_bootstrap}
\title{Bootstrapping}
\usage{
set_bootstrap(
  num_boot = 1000,
  boot_factor = 1000,
  file = here("inst", "Example1.BSN")
)
}
\arguments{
\item{num_boot}{Number Of Bootstrap Operations}

\item{boot_factor}{Population scaling Factor}

\item{file}{Bootstrap File}
}
\description{
AGEPRO Bootstrapping options
}
