% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_input.R
\name{set_general}
\alias{set_general}
\title{AGEPRO general options}
\usage{
set_general(
  first_year,
  last_year,
  first_age,
  last_age,
  num_sims,
  num_fleets,
  num_models,
  seed,
  discard = FALSE
)
}
\arguments{
\item{first_year}{First Year of prediction}

\item{last_year}{Last Year of prediction}

\item{first_age}{First Age Class}

\item{last_age}{Last Age Class}

\item{num_sims}{Number of Population Simulations}

\item{num_fleets}{Number of Fleets}

\item{num_models}{Number of Recruitment Models}

\item{seed}{Psuedo-random number seed}

\item{discard}{Are Discards Are Present}
}
\description{
Sets general options for AGEPRO
}
