% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model.R
\name{set_job_name}
\alias{set_job_name}
\title{Job directory name setup}
\usage{
set_job_name(model, type = c("filename", "case_id"))
}
\arguments{
\item{model}{Agepro model}

\item{type}{Option to name to use input model filename or the model's
case_id as part of the job_dir name. Defaults to "filename"}
}
\value{
Returns the Character String, the contents dependent on the "type"
requested. Using "case_id" returns the case id model name formatted
for file directories. NULL or blank "filename" values return \code{untitled},
otherwise function will return the input "filename" basename without
the file extension.
}
\description{
Helper method that uses the agepro model's data to label the job directory
names.
}
