% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recruitment.R
\name{set_recruitment}
\alias{set_recruitment}
\title{set recruitment}
\usage{
set_recruitment(model, recFac, ssbFac, maxRecObs, prob)
}
\arguments{
\item{model}{Recruitment model}

\item{recFac}{Recruitment Scaling Factor}

\item{ssbFac}{Spawning Stock Biomass (SSB) Scaling Factor}

\item{maxRecObs}{Limits on Maximun Records Observed}

\item{prob}{Recruitment Probability over each time period}
}
\description{
Writes the generalized AGEPRO Recruitment json data object.
}
\details{
TODO: Recruitment models will call this function.
}
