% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_input.R
\name{set_stochastic_age}
\alias{set_stochastic_age}
\title{Sets Stochastic age}
\usage{
set_stochastic_age(valid_flag = c(1, 2))
}
\arguments{
\item{valid_flag}{Valid flag}
}
\description{
Sets a Stochastic parameter of Age
}
