% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_input.R
\name{set_stochastic_weight_age}
\alias{set_stochastic_weight_age}
\title{Stochastic (Weight) Of Age}
\usage{
set_stochastic_weight_age(df, sd, flag = 1, time_flag = TRUE)
}
\arguments{
\item{df}{data}

\item{sd}{error}

\item{flag}{flag}

\item{time_flag}{Time Varying}
}
\description{
Sets a Stochastic (Weight) of Age
}
