% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_input.R
\name{set_stochasticWeightAge}
\alias{set_stochasticWeightAge}
\title{Stochastic (Weight) Of Age}
\usage{
set_stochasticWeightAge(df, sd, flag = 1, timeFlag = TRUE)
}
\arguments{
\item{df}{data}

\item{sd}{error}

\item{flag}{flag}

\item{timeFlag}{Time Varying}
}
\description{
Sets a Stochastic (Weight) of Age
}
