% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeInput.R
\name{set_version}
\alias{set_version}
\title{AGEPRO version}
\usage{
set_version(
  legacyVersionString = "AGEPRO VERSION 4.0",
  versionNumber = "4.0.0.0"
)
}
\arguments{
\item{legacyVersionString}{Version Character String generated by legacy versions}

\item{versionNumber}{AGEPRO version}
}
\description{
Sets/marks AGEPRO Calcuation Version
}
