% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_agepro_model.R
\name{setup_agepro_model}
\alias{setup_agepro_model}
\title{Function wrapper to create a agepro_model}
\usage{
setup_agepro_model(type = c("inp", "json"), ...)
}
\arguments{
\item{type}{Create a agepro_inp_model (default), or agpepro_json_model}

\item{...}{further arguments passed to or from other methods}
}
\value{
R6class ojject cotaining the agepro_model class object
}
\description{
Functcion wrapper to create agepro_model R6class. This function retruns a
new instance of the agepro_inp_model or agepro_json_model
}
\examples{
\dontrun{
# General parameters for 2019-2026 Uku Projections Base (Example 4)
test <- setup_agepro_model(
  type="inp",
  yr_start=2019,
  yr_end= 2026,
  age_begin = 1,
  age_end = 32,
  num_pop_sims = 1000,
  num_fleets = 4,
  num_rec_models = 1,
  discards_present = 0,
  seed = 300)
}

}
