% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruit_model.R
\name{shepherd_curve_model}
\alias{shepherd_curve_model}
\title{Shepherd Curve with Lognormal Error (Model #7)}
\description{
Shepherd Curve with Lognormal Error (Model #7)
}
\section{Super classes}{
\code{\link[ageproR:recruit_model]{recruit_model}} -> \code{\link[ageproR:parametric_curve]{parametric_curve}} -> \code{shepherd_curve_model}
}
\section{Active bindings}{
  \if{html}{\out{<div class="r6-active-bindings">}}
  \describe{
    \item{\code{kpar}}{\cr
Sets the Parametric Curve Parameter, k. Returns the
current value if no argument was passed}

    \item{\code{json_recruit_data}}{Returns JSON-ready Recruit Model Data}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-shepherd_curve_model-initialize}{\code{shepherd_curve_model$new()}}
    \item \href{#method-shepherd_curve_model-print}{\code{shepherd_curve_model$print()}}
    \item \href{#method-shepherd_curve_model-read_inp_lines}{\code{shepherd_curve_model$read_inp_lines()}}
    \item \href{#method-shepherd_curve_model-inp_lines_recruit_data}{\code{shepherd_curve_model$inp_lines_recruit_data()}}
    \item \href{#method-shepherd_curve_model-clone}{\code{shepherd_curve_model$clone()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-shepherd_curve_model-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-shepherd_curve_model-initialize}{}}}
\subsection{\code{shepherd_curve_model$new()}}{
  Initializes the Shepherd Curve Model
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{shepherd_curve_model$new(alpha = 0.1, beta = 0.1, kpar = 0.1, variance = 0.1)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{alpha}}{\cr
Stock Recruitment Parameter, alpha}
      \item{\code{beta}}{\cr
Stock Recruitment Parameter, beta}
      \item{\code{kpar}}{kpar}
      \item{\code{variance}}{\cr
variance}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-shepherd_curve_model-print"></a>}}
\if{latex}{\out{\hypertarget{method-shepherd_curve_model-print}{}}}
\subsection{\code{shepherd_curve_model$print()}}{
  Prints out Parametric Curve Data
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{shepherd_curve_model$print(...)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{...}}{further arguments passed to or from other methods}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-shepherd_curve_model-read_inp_lines"></a>}}
\if{latex}{\out{\hypertarget{method-shepherd_curve_model-read_inp_lines}{}}}
\subsection{\code{shepherd_curve_model$read_inp_lines()}}{
  Reads Parametric Curve model data from AGEPRO Input file
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{shepherd_curve_model$read_inp_lines(inp_con, nline)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{inp_con}}{Open file connection to AGEPRO Input File.}
      \item{\code{nline}}{\cr
Location of the current line number being read when the file connection to
the AGEPRO input file is open.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-shepherd_curve_model-inp_lines_recruit_data"></a>}}
\if{latex}{\out{\hypertarget{method-shepherd_curve_model-inp_lines_recruit_data}{}}}
\subsection{\code{shepherd_curve_model$inp_lines_recruit_data()}}{
  Exports RECRUIT submodel data for shepherd curve recruitment
to AGEPRO input file lines.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{shepherd_curve_model$inp_lines_recruit_data(delimiter = " ")}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{delimiter}}{Character string delimiter separating values of AGEPRO input file line.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-shepherd_curve_model-clone"></a>}}
\if{latex}{\out{\hypertarget{method-shepherd_curve_model-clone}{}}}
\subsection{\code{shepherd_curve_model$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{shepherd_curve_model$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
