% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecruitModel.R
\name{ShepherdCurveModel}
\alias{ShepherdCurveModel}
\title{Shepherd Curve with Lognormal Error (Model #7)}
\description{
Shepherd Curve with Lognormal Error (Model #7)

Shepherd Curve with Lognormal Error (Model #7)
}
\section{Super classes}{
\code{\link[AGEPROjson:RecruitModel]{AGEPROjson::RecruitModel}} -> \code{AGEPROjson::ParametricRecruitModel} -> \code{ShepherdCurveModel}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{kpar}}{Stock Recruitment Parameter, k}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{set_kpar}}{\cr
Sets the Parametric Curve Parameter, k. Returns the
current value if no argument was passed}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ShepherdCurveModel-new}{\code{ShepherdCurveModel$new()}}
\item \href{#method-ShepherdCurveModel-print}{\code{ShepherdCurveModel$print()}}
\item \href{#method-ShepherdCurveModel-clone}{\code{ShepherdCurveModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ShepherdCurveModel-new"></a>}}
\if{latex}{\out{\hypertarget{method-ShepherdCurveModel-new}{}}}
\subsection{Method \code{new()}}{
Initializes the Shepherd Curve Model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShepherdCurveModel$new(alpha = 0, beta = 0, kpar = 0, variance = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{alpha}}{\cr
Stock Recruitment Parameter, alpha}

\item{\code{beta}}{\cr
Stock Recruitment Parameter, beta}

\item{\code{kpar}}{kpar}

\item{\code{variance}}{\cr
variance}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ShepherdCurveModel-print"></a>}}
\if{latex}{\out{\hypertarget{method-ShepherdCurveModel-print}{}}}
\subsection{Method \code{print()}}{
Prints out Parametric Curve Data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShepherdCurveModel$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{further arguments passed to or from other methods}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ShepherdCurveModel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ShepherdCurveModel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShepherdCurveModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
