% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochastic.R
\name{stochastic}
\alias{stochastic}
\title{Stochastic AGEPRO Keyword Parameter Structure.}
\description{
Generalized Class Structure for Stochastic AGEPRO Keyword parameters.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{input_option}}{Stochastic Input option}

\item{\code{time_varying}}{Logical flag to list stochastic data by observation
year}

\item{\code{stochastic_table}}{This is the logic for the average stochastic
AGEPRO keyword parameter's at age (and by fleet if fleets are a
factor).}

\item{\code{cv_table}}{Matrix containing the vector of of age-specific CVs for
sampling the average stochastic AGEPRO keyword parameter's at age
(and by fleet if fleets are a factor) with lognormal process error.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-stochastic-new}{\code{stochastic$new()}}
\item \href{#method-stochastic-create_stochastic_table}{\code{stochastic$create_stochastic_table()}}
\item \href{#method-stochastic-clone}{\code{stochastic$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-stochastic-new"></a>}}
\if{latex}{\out{\hypertarget{method-stochastic-new}{}}}
\subsection{Method \code{new()}}{
Initializes the stochastic class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{stochastic$new(
  num_obs_yrs,
  num_ages,
  num_fleets = 1,
  input_option = 1,
  time_flag = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{num_obs_yrs}}{Numbers of years in from first to last year of
projection.}

\item{\code{num_ages}}{Number of Age classes}

\item{\code{num_fleets}}{Number of Fleets. Default is 1}

\item{\code{time_flag}}{Logical flag that enables the stochastic parameter to use
as a time-varying array if TRUE (or 1). Otherwise, FALSE the vector will
cover "all years" of the projection. Default is TRUE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-stochastic-create_stochastic_table"></a>}}
\if{latex}{\out{\hypertarget{method-stochastic-create_stochastic_table}{}}}
\subsection{Method \code{create_stochastic_table()}}{
Creates an stochastic table
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{stochastic$create_stochastic_table(num_obs_yrs, num_ages, num_fleets = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{num_obs_yrs}}{Numbers of years in from first to last year of
projection.}

\item{\code{num_ages}}{Number of Age classes}

\item{\code{num_fleets}}{Number of Fleets. Default is 1}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-stochastic-clone"></a>}}
\if{latex}{\out{\hypertarget{method-stochastic-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{stochastic$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
