% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_percentile_summary.R
\name{user_percentile_summary}
\alias{user_percentile_summary}
\title{Percentile summary of the key results of AGEPRO projection output}
\description{
Class Structure that includes user-defined options for setting
a specific percentile for the distributions of outputs.
}
\section{Super class}{
\code{\link[ageproR:agepro_options_flags]{ageproR::agepro_options_flags}} -> \code{user_percentile_summary}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{report_percentile}}{User-defined percentile for reporting the percentile of the projected
distribution of the following quantities of interest by year:
spawning stock biomass, stock biomass on January 1st, mean biomass,
combined catch biomass, landings, fishing mortality, and stock
numbers at age}

\item{\code{keyword_name}}{AGEPRO keyword parameter name}

\item{\code{inp_keyword}}{Returns AGEPRO input-file formatted Parameter}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-user_percentile_summary-new}{\code{user_percentile_summary$new()}}
\item \href{#method-user_percentile_summary-clone}{\code{user_percentile_summary$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-user_percentile_summary-new"></a>}}
\if{latex}{\out{\hypertarget{method-user_percentile_summary-new}{}}}
\subsection{Method \code{new()}}{
Initialize the class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{user_percentile_summary$new(perc = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{perc}}{User-defined percentile of projected distributions}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-user_percentile_summary-clone"></a>}}
\if{latex}{\out{\hypertarget{method-user_percentile_summary-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{user_percentile_summary$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
