% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_logical_parameter}
\alias{validate_logical_parameter}
\title{Validate parameters formatted as logical values}
\usage{
validate_logical_parameter(x)
}
\arguments{
\item{x}{Object to Check}
}
\description{
Generalized validation method to check input value of parameter formatted
as a logical. AGEPRO, and its input file format, reads logical values as
\code{0} (FALSE) and  \code{1} (TRUE). In R, these numeric values can be interpreted
as logical values.
}
