% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_map}
\alias{validate_map}
\title{Custom mapping function for error handing}
\usage{
validate_map(.xs, .fn, ...)
}
\arguments{
\item{.xs}{List r Atomic Vector}

\item{.fn}{Function}

\item{...}{further arguments passed to or from other methods}
}
\description{
Custom mapping function used for error handling. This is based on the
rlang topic errors guide.
}
