% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_bsn_from_existing_file}
\alias{write_bsn_from_existing_file}
\title{Import existing bootstrap file to create a new one that matches the
dimensions for a AGEPRO model.}
\usage{
write_bsn_from_existing_file(
  bsn_file,
  out_file,
  target_ages,
  target_runs = 1000
)
}
\arguments{
\item{bsn_file}{Input Bootstrap Filename}

\item{out_file}{Output Formatted Bootstrap Filename}

\item{target_ages}{Target Number of Ages (Columns) of formatted output
bootstrap file}

\item{target_runs}{Target Number of Bootstrap Runs (Rows) of formatted output
bootstrap file}
}
\description{
Note: This function is intended to test Bootstrap compatability with the
Agepro Input File model and is not intended for production use.
}
