% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model.R
\name{write_logfile}
\alias{write_logfile}
\title{Log AGEPRO Ouput to File}
\usage{
write_logfile(cout, fn_logfile)
}
\arguments{
\item{cout}{Output connection from console}

\item{fn_logfile}{Filename path to save output}
}
\description{
Saves AGEPRO Output to Logfile
}
