% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agepro_model.R
\name{agepro_inp_model}
\alias{agepro_inp_model}
\title{AGEPRO Input File Model}
\description{
AGEPRO Input File Model

AGEPRO Input File Model
}
\details{
File Functionality is based on AGEPRO-CoreLib implementation
}
\section{Super class}{
\code{\link[ageproj:agepro_model]{ageproj::agepro_model}} -> \code{agepro_inp_model}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{nline}}{nlines}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-agepro_inp_model-new}{\code{agepro_inp_model$new()}}
\item \href{#method-agepro_inp_model-read_inp}{\code{agepro_inp_model$read_inp()}}
\item \href{#method-agepro_inp_model-read_inpfile_values}{\code{agepro_inp_model$read_inpfile_values()}}
\item \href{#method-agepro_inp_model-match_keyword}{\code{agepro_inp_model$match_keyword()}}
\item \href{#method-agepro_inp_model-assert_inpfile_version}{\code{agepro_inp_model$assert_inpfile_version()}}
\item \href{#method-agepro_inp_model-not_implemented}{\code{agepro_inp_model$not_implemented()}}
\item \href{#method-agepro_inp_model-write_inp}{\code{agepro_inp_model$write_inp()}}
\item \href{#method-agepro_inp_model-clone}{\code{agepro_inp_model$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="ageproj" data-topic="agepro_model" data-id="set_bootstrap_filename"><a href='../../ageproj/html/agepro_model.html#method-agepro_model-set_bootstrap_filename'><code>ageproj::agepro_model$set_bootstrap_filename()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ageproj" data-topic="agepro_model" data-id="set_recruit_model"><a href='../../ageproj/html/agepro_model.html#method-agepro_model-set_recruit_model'><code>ageproj::agepro_model$set_recruit_model()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-agepro_inp_model-new"></a>}}
\if{latex}{\out{\hypertarget{method-agepro_inp_model-new}{}}}
\subsection{Method \code{new()}}{
Initializes the input file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{agepro_inp_model$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-agepro_inp_model-read_inp"></a>}}
\if{latex}{\out{\hypertarget{method-agepro_inp_model-read_inp}{}}}
\subsection{Method \code{read_inp()}}{
Read AGEPRO INP Input Files
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{agepro_inp_model$read_inp(inpfile)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{inpfile}}{input file name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-agepro_inp_model-read_inpfile_values"></a>}}
\if{latex}{\out{\hypertarget{method-agepro_inp_model-read_inpfile_values}{}}}
\subsection{Method \code{read_inpfile_values()}}{
Read Input file Values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{agepro_inp_model$read_inpfile_values(inp_con)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{inp_con}}{Open file connection to AGEPRO Input File.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-agepro_inp_model-match_keyword"></a>}}
\if{latex}{\out{\hypertarget{method-agepro_inp_model-match_keyword}{}}}
\subsection{Method \code{match_keyword()}}{
Match Keyword

TODO: ~~CASEID~~, ~~GENERAL~~, RECRUIT, STOCK_WEIGHT, SSB_WEIGHT,
MEAN_WEIGHT, CATCH_WEIGHT, DISC_WEIGHT, NATMORT, MATURITY,
FISHERY, DISCARD, BIOLOGICAL, ~~BOOTSTRAP~~, HARVEST, REBUILD
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{agepro_inp_model$match_keyword(inp_line, inp_con)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{inp_line}}{Line read from the AGEPRO Input File}

\item{\code{inp_con}}{Open file connection to AGEPRO Input File.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-agepro_inp_model-assert_inpfile_version"></a>}}
\if{latex}{\out{\hypertarget{method-agepro_inp_model-assert_inpfile_version}{}}}
\subsection{Method \code{assert_inpfile_version()}}{
Check Input File Version
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{agepro_inp_model$assert_inpfile_version(inp_line)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{inp_line}}{Line read from the AGEPRO Input File}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-agepro_inp_model-not_implemented"></a>}}
\if{latex}{\out{\hypertarget{method-agepro_inp_model-not_implemented}{}}}
\subsection{Method \code{not_implemented()}}{
Throws a Not Implemented exception message. Placeholder function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{agepro_inp_model$not_implemented(keyword = "")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{keyword}}{keyword}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-agepro_inp_model-write_inp"></a>}}
\if{latex}{\out{\hypertarget{method-agepro_inp_model-write_inp}{}}}
\subsection{Method \code{write_inp()}}{
Writes AGEPRO keyword parameter data as a AGEPRO input file (*.inp)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{agepro_inp_model$write_inp(inpfile)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{inpfile}}{input file path}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-agepro_inp_model-clone"></a>}}
\if{latex}{\out{\hypertarget{method-agepro_inp_model-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{agepro_inp_model$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
