% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agepro_model.R
\name{agepro_json_model}
\alias{agepro_json_model}
\title{AGEPRO model w/ JSON input file bindings}
\description{
File Functionality on experimental JSON input file
}
\section{Super class}{
\code{\link[ageproR:agepro_model]{ageproR::agepro_model}} -> \code{agepro_json_model}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-agepro_json_model-get_json}{\code{agepro_json_model$get_json()}}
\item \href{#method-agepro_json_model-write_json}{\code{agepro_json_model$write_json()}}
\item \href{#method-agepro_json_model-clone}{\code{agepro_json_model$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="ageproR" data-topic="agepro_model" data-id="initialize"><a href='../../ageproR/html/agepro_model.html#method-agepro_model-initialize'><code>ageproR::agepro_model$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ageproR" data-topic="agepro_model" data-id="set_bootstrap_filename"><a href='../../ageproR/html/agepro_model.html#method-agepro_model-set_bootstrap_filename'><code>ageproR::agepro_model$set_bootstrap_filename()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ageproR" data-topic="agepro_model" data-id="set_recruit_model"><a href='../../ageproR/html/agepro_model.html#method-agepro_model-set_recruit_model'><code>ageproR::agepro_model$set_recruit_model()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-agepro_json_model-get_json"></a>}}
\if{latex}{\out{\hypertarget{method-agepro_json_model-get_json}{}}}
\subsection{Method \code{get_json()}}{
Get json
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{agepro_json_model$get_json()}\if{html}{\out{</div>}}
}

\subsection{Details}{
See \link[jsonlite:fromJSON]{jsonlite::toJSON} for more details. \code{NULL} values is written as
\code{{}}. \code{NA} values is written as \code{null}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-agepro_json_model-write_json"></a>}}
\if{latex}{\out{\hypertarget{method-agepro_json_model-write_json}{}}}
\subsection{Method \code{write_json()}}{
Write JSON file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{agepro_json_model$write_json(show_dir = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{show_dir}}{Option to show directory after JSON file is written.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-agepro_json_model-clone"></a>}}
\if{latex}{\out{\hypertarget{method-agepro_json_model-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{agepro_json_model$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
