% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agepro_model.R
\name{agepro_model}
\alias{agepro_model}
\title{R6 class representing AGEPRO model}
\description{
AGEPRO model contains the projection time horizon, age class range, number
of fleets, recruitment, and unertainties
}
\details{
AGEPRO performs stochastic projections on exploited fisheries stock to
determine age-structured population over a time period. Brodziak, 2022
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{recruit}}{AGEPRO Recruitmment Model(s)}

\item{\code{case_id}}{Case id}

\item{\code{bootstrap}}{Bootstrapping}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{ver_legacy_string}}{Version string on AGEPRO input files (*.inp) for version compatibility
with Jon Brodiak's AGEPRO calculation engine.}

\item{\code{ver_numeric_string}}{Numeric character string based by Semantic-like versioning format.}

\item{\code{general}}{General Options}

\item{\code{natmort}}{Natural Mortality}

\item{\code{maturity}}{Maturity Fraction}

\item{\code{fishery}}{\cr
Fishery Selectivity}

\item{\code{discard}}{\cr
Discard Fraction}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-agepro_model-new}{\code{agepro_model$new()}}
\item \href{#method-agepro_model-set_recruit_model}{\code{agepro_model$set_recruit_model()}}
\item \href{#method-agepro_model-set_bootstrap_filename}{\code{agepro_model$set_bootstrap_filename()}}
\item \href{#method-agepro_model-clone}{\code{agepro_model$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-agepro_model-new"></a>}}
\if{latex}{\out{\hypertarget{method-agepro_model-new}{}}}
\subsection{Method \code{new()}}{
Starts an instances of the AGEPRO Model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{agepro_model$new(
  yr_start,
  yr_end,
  age_begin,
  age_end,
  num_pop_sims,
  num_fleets,
  num_rec_models,
  discards = FALSE,
  seed = sample.int(1e+08, 1)
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{yr_start}}{First Year of Projection}

\item{\code{yr_end}}{Last Year of Projection}

\item{\code{age_begin}}{Age begin}

\item{\code{age_end}}{Age end}

\item{\code{num_pop_sims}}{Number of population simulations}

\item{\code{num_fleets}}{Number of fleets}

\item{\code{num_rec_models}}{Number of Recruit Modules}

\item{\code{discards}}{discards. FALSE by default}

\item{\code{seed}}{Random Number seed. A pesdorandom number is set as default.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-agepro_model-set_recruit_model"></a>}}
\if{latex}{\out{\hypertarget{method-agepro_model-set_recruit_model}{}}}
\subsection{Method \code{set_recruit_model()}}{
Set model's Recruitment model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{agepro_model$set_recruit_model(model_num)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model_num}}{\cr
AGEPRO Recruitment Model Number}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-agepro_model-set_bootstrap_filename"></a>}}
\if{latex}{\out{\hypertarget{method-agepro_model-set_bootstrap_filename}{}}}
\subsection{Method \code{set_bootstrap_filename()}}{
Wrapper function to call bootstrap's set_bootstrap_filename
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{agepro_model$set_bootstrap_filename(bsnfile)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{bsnfile}}{bootstrap filename}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-agepro_model-clone"></a>}}
\if{latex}{\out{\hypertarget{method-agepro_model-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{agepro_model$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
