% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agepro_options_flags.R
\name{agepro_options_flags}
\alias{agepro_options_flags}
\title{Class container that encapsulates AGEPRO user-defined options}
\description{
Class Structure that encapsulates AGEPRO's user-defined options including
output options, percentile summary, scaling factors, biological reference
points, maximum bounds, and retrospective adjustment.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{enable_user_percentile_summary}}{\link[base:logical]{Logical} flag to allow percentile summary of the key
results in the output file.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-agepro_options_flags-new}{\code{agepro_options_flags$new()}}
\item \href{#method-agepro_options_flags-set_enable_user_percentile_summary}{\code{agepro_options_flags$set_enable_user_percentile_summary()}}
\item \href{#method-agepro_options_flags-clone}{\code{agepro_options_flags$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-agepro_options_flags-new"></a>}}
\if{latex}{\out{\hypertarget{method-agepro_options_flags-new}{}}}
\subsection{Method \code{new()}}{
Initialize the class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{agepro_options_flags$new(enable_percentile = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{enable_percentile}}{\link[base:logical]{Logical} flag to enable percentile summary of the key
results in the output file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-agepro_options_flags-set_enable_user_percentile_summary"></a>}}
\if{latex}{\out{\hypertarget{method-agepro_options_flags-set_enable_user_percentile_summary}{}}}
\subsection{Method \code{set_enable_user_percentile_summary()}}{
Toggles the logical flag to enable user_percentile summary
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{agepro_options_flags$set_enable_user_percentile_summary(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{Logical value}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-agepro_options_flags-clone"></a>}}
\if{latex}{\out{\hypertarget{method-agepro_options_flags-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{agepro_options_flags$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
