% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AgeproModel.R
\name{AgeproModel}
\alias{AgeproModel}
\title{R6 class representing AGEPRO model}
\description{
AGEPRO model contains the projection time horizon, age class range, number
of fleets, recruitment, and unertainties
}
\details{
AGEPRO performs stochastic projections on exploited fisheries stock to
determine age-structured population over a time period. Brodziak, 2022
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{general}}{General Parameters}

\item{\code{recruit}}{AGEPRO Recruitmment Model(s)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AgeproModel-new}{\code{AgeproModel$new()}}
\item \href{#method-AgeproModel-set_recruit_model}{\code{AgeproModel$set_recruit_model()}}
\item \href{#method-AgeproModel-get_json}{\code{AgeproModel$get_json()}}
\item \href{#method-AgeproModel-clone}{\code{AgeproModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AgeproModel-new"></a>}}
\if{latex}{\out{\hypertarget{method-AgeproModel-new}{}}}
\subsection{Method \code{new()}}{
Starts an instances of the AGEPRO Model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AgeproModel$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Parameters to pass to GeneralParams}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AgeproModel-set_recruit_model"></a>}}
\if{latex}{\out{\hypertarget{method-AgeproModel-set_recruit_model}{}}}
\subsection{Method \code{set_recruit_model()}}{
Set model's Recruitment model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AgeproModel$set_recruit_model(model_num)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model_num}}{\cr
AGEPRO Recruitment Model Number}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AgeproModel-get_json"></a>}}
\if{latex}{\out{\hypertarget{method-AgeproModel-get_json}{}}}
\subsection{Method \code{get_json()}}{
Get json
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AgeproModel$get_json()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AgeproModel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AgeproModel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AgeproModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
