% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AgeproModel.R
\name{AgeproModel}
\alias{AgeproModel}
\title{R6 class representing AGEPRO model}
\description{
AGEPRO model contains the projection time horizon, age class range, number
of fleets, recruitment, and unertainties
}
\details{
AGEPRO performs stochastic projections on exploited fisheries stock to
determine age-structured population over a time period. Brodziak, 2022
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{general}}{General Parameters}

\item{\code{recruit}}{AGEPRO Recruitmment Model(s)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AgeproModel-new}{\code{AgeproModel$new()}}
\item \href{#method-AgeproModel-set_recruit_model}{\code{AgeproModel$set_recruit_model()}}
\item \href{#method-AgeproModel-get_json}{\code{AgeproModel$get_json()}}
\item \href{#method-AgeproModel-write_json}{\code{AgeproModel$write_json()}}
\item \href{#method-AgeproModel-clone}{\code{AgeproModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AgeproModel-new"></a>}}
\if{latex}{\out{\hypertarget{method-AgeproModel-new}{}}}
\subsection{Method \code{new()}}{
Starts an instances of the AGEPRO Model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AgeproModel$new(
  yr_start,
  yr_end,
  age_begin,
  age_end,
  num_fleets,
  num_rec_models,
  num_pop_sims,
  discards = FALSE,
  seed = sample.int(1e+08, 1)
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{yr_start}}{First Year of Projection}

\item{\code{yr_end}}{Last Year of Projection}

\item{\code{age_begin}}{Age begin}

\item{\code{age_end}}{Age end}

\item{\code{num_fleets}}{Number of fleets}

\item{\code{num_rec_models}}{Number of Recruit Modules}

\item{\code{num_pop_sims}}{Number of population simuations}

\item{\code{discards}}{discards. FALSE by default}

\item{\code{seed}}{Random Number seed. A pesdorandom number is set as default.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AgeproModel-set_recruit_model"></a>}}
\if{latex}{\out{\hypertarget{method-AgeproModel-set_recruit_model}{}}}
\subsection{Method \code{set_recruit_model()}}{
Set model's Recruitment model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AgeproModel$set_recruit_model(model_num)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model_num}}{\cr
AGEPRO Recruitment Model Number}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AgeproModel-get_json"></a>}}
\if{latex}{\out{\hypertarget{method-AgeproModel-get_json}{}}}
\subsection{Method \code{get_json()}}{
Get json
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AgeproModel$get_json()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AgeproModel-write_json"></a>}}
\if{latex}{\out{\hypertarget{method-AgeproModel-write_json}{}}}
\subsection{Method \code{write_json()}}{
Write JSON file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AgeproModel$write_json()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AgeproModel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AgeproModel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AgeproModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
