% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ageproR-package.R
\docType{package}
\name{ageproR-package}
\alias{ageproR}
\alias{ageproR-package}
\title{ageproR: Age Structured Projection Model (AGEPRO) data structure}
\description{
Object-oriented data structures based on the input file format defined by the Age Structured Projection Model Reference Manual. Includes support modules for AGEPRO input file support and interactivity with AGEPRO calculation engine.
}
\author{
\strong{Maintainer}: Eric Fletcher \email{eric.fletcher@noaa.gov}

Authors:
\itemize{
  \item Jon Brodziak
}

}
\keyword{internal}
