% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ageproWrapper.R
\name{ageproWrapper}
\alias{ageproWrapper}
\title{R6Class Wrapper for Running AGEPRO Calcuation Engine binary}
\description{
A R6class wrapper that handles exeution of the AGEPRO calcuation binary.
This includes varaibles to handle binary path and output logfile, and
a function wrapper to run the binary.
}
\section{Active bindings}{
  \if{html}{\out{<div class="r6-active-bindings">}}
  \describe{
    \item{\code{agepro_path}}{AGEPRO Calculation Engine path}

    \item{\code{logdir}}{File path reserved for logfiles for AGEPRO Calcuation Engine runs.}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-ageproWrapper-initialize}{\code{ageproWrapper$new()}}
    \item \href{#method-ageproWrapper-run}{\code{ageproWrapper$run()}}
    \item \href{#method-ageproWrapper-run_logfile}{\code{ageproWrapper$run_logfile()}}
    \item \href{#method-ageproWrapper-validate_binary}{\code{ageproWrapper$validate_binary()}}
    \item \href{#method-ageproWrapper-clone}{\code{ageproWrapper$clone()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ageproWrapper-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-ageproWrapper-initialize}{}}}
\subsection{\code{ageproWrapper$new()}}{
  Initializes the binary Wrapper
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ageproWrapper$new(path = NULL, warn = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{path}}{Explicit path to directory}
      \item{\code{warn}}{Logical parameter to validate path at initalization. If enabled,
invalid reuslts returns a warning.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ageproWrapper-run"></a>}}
\if{latex}{\out{\hypertarget{method-ageproWrapper-run}{}}}
\subsection{\code{ageproWrapper$run()}}{
  Runs the Binary
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ageproWrapper$run(args = character())}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{args}}{Program agruments}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ageproWrapper-run_logfile"></a>}}
\if{latex}{\out{\hypertarget{method-ageproWrapper-run_logfile}{}}}
\subsection{\code{ageproWrapper$run_logfile()}}{
  Runs the Binary with logfile

This function uses the private function get_logdir
to return the logfile directory. This function will return
the custom output directory if it exists. Otherwise it will
check if user session (.Rprofile) has an existing path. If
\code{ageproR.run_logdir} does't work it, will fallback to the "AGEPRO"
subirectory of home directory.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ageproWrapper$run_logfile(args = character(), out_dir = "")}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{args}}{Program arguments}
      \item{\code{out_dir}}{Output directory for the logfile. The default
blank character string will return the AGEPRO subdirectory
of the home directory via get_logdir

Validate file path for AGEPRO Calcuation Engine Binary}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ageproWrapper-validate_binary"></a>}}
\if{latex}{\out{\hypertarget{method-ageproWrapper-validate_binary}{}}}
\subsection{\code{ageproWrapper$validate_binary()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ageproWrapper$validate_binary(exepath = file.path(getwd(), "agepro.exe"))}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{exepath}}{AGEPRO Calcuuation Path Binary}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ageproWrapper-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ageproWrapper-clone}{}}}
\subsection{\code{ageproWrapper$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ageproWrapper$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
