% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ageproWrapper.R
\name{ageproWrapper}
\alias{ageproWrapper}
\title{ageproWrapper}
\description{
ageproWrapper

ageproWrapper
}
\details{
ageproWrapper
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{agepro_path}}{AGEPRO Calculation Engine path}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ageproWrapper-new}{\code{ageproWrapper$new()}}
\item \href{#method-ageproWrapper-run}{\code{ageproWrapper$run()}}
\item \href{#method-ageproWrapper-run_logfile}{\code{ageproWrapper$run_logfile()}}
\item \href{#method-ageproWrapper-clone}{\code{ageproWrapper$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ageproWrapper-new"></a>}}
\if{latex}{\out{\hypertarget{method-ageproWrapper-new}{}}}
\subsection{Method \code{new()}}{
Initializes the binary Wrapper
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ageproWrapper$new(path = NULL, warn = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{Explicit path to directory}

\item{\code{warn}}{Logical parameter to validate path at initalization. If enabled,
invalid reuslts returns a warning.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ageproWrapper-run"></a>}}
\if{latex}{\out{\hypertarget{method-ageproWrapper-run}{}}}
\subsection{Method \code{run()}}{
Runs the Binary
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ageproWrapper$run(args = character())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{args}}{Program agruments}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ageproWrapper-run_logfile"></a>}}
\if{latex}{\out{\hypertarget{method-ageproWrapper-run_logfile}{}}}
\subsection{Method \code{run_logfile()}}{
Runs the Binary with logfile
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ageproWrapper$run_logfile(args = character())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{args}}{Program arguments}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ageproWrapper-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ageproWrapper-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ageproWrapper$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
