% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruit_model.R
\name{beverton_holt_curve_model}
\alias{beverton_holt_curve_model}
\title{Beverton-Holt w/ Lognormal Error (Model #5)}
\description{
Beverton-Holt w/ Lognormal Error (Model #5)

Beverton-Holt w/ Lognormal Error (Model #5)
}
\section{Super classes}{
\code{\link[ageproj:recruit_model]{ageproj::recruit_model}} -> \code{\link[ageproj:parametric_curve]{ageproj::parametric_curve}} -> \code{beverton_holt_curve_model}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-beverton_holt_curve_model-new}{\code{beverton_holt_curve_model$new()}}
\item \href{#method-beverton_holt_curve_model-clone}{\code{beverton_holt_curve_model$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="ageproj" data-topic="parametric_curve" data-id="print"><a href='../../ageproj/html/parametric_curve.html#method-parametric_curve-print'><code>ageproj::parametric_curve$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-beverton_holt_curve_model-new"></a>}}
\if{latex}{\out{\hypertarget{method-beverton_holt_curve_model-new}{}}}
\subsection{Method \code{new()}}{
Initializes the Beverton Holt Curve Model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{beverton_holt_curve_model$new(alpha = 0, beta = 0, variance = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{alpha}}{\cr
Stock Recruitment Parameter, alpha}

\item{\code{beta}}{\cr
Stock Recruitment Parameter, beta}

\item{\code{variance}}{\cr
variance}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-beverton_holt_curve_model-clone"></a>}}
\if{latex}{\out{\hypertarget{method-beverton_holt_curve_model-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{beverton_holt_curve_model$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
