% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap}
\alias{bootstrap}
\title{Input information for bootstrap numbers at age file}
\description{
-The Number of data values of each row must equal to the number of age classes.
-The number of rows in a bootstrap file must be at least equal to the number of bootstrap
iterations containing the popluation of the first year in the projection
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{num_bootstraps}}{Number of bootstraps replicates of initial popualion size}

\item{\code{pop_scale_factor}}{Population Scale Factor, or BootFac, that represents the multiplicative
factor to convert the relative bootstrap population numbers at age to
absolute numbers at age.}

\item{\code{bootstrap_file}}{Bootstrap file path.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-bootstrap-new}{\code{bootstrap$new()}}
\item \href{#method-bootstrap-clone}{\code{bootstrap$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-bootstrap-new"></a>}}
\if{latex}{\out{\hypertarget{method-bootstrap-new}{}}}
\subsection{Method \code{new()}}{
Initializes the Bootstrap Class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{bootstrap$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-bootstrap-clone"></a>}}
\if{latex}{\out{\hypertarget{method-bootstrap-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{bootstrap$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
