% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_id.R
\name{case_id}
\alias{case_id}
\title{AGEPRO case id}
\description{
AGEPRO case id

AGEPRO case id
}
\details{
Input title identifying model attributes
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{case_id}}{case id}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-case_id-read_inp_lines}{\code{case_id$read_inp_lines()}}
\item \href{#method-case_id-clone}{\code{case_id$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-case_id-read_inp_lines"></a>}}
\if{latex}{\out{\hypertarget{method-case_id-read_inp_lines}{}}}
\subsection{Method \code{read_inp_lines()}}{
Read AGEPRO Case ID from input data file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{case_id$read_inp_lines(con, nline)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{con}}{input file connection}

\item{\code{nline}}{Line number}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-case_id-clone"></a>}}
\if{latex}{\out{\hypertarget{method-case_id-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{case_id$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
