% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{check_agepro_model_class}
\alias{check_agepro_model_class}
\alias{assert_agepro_model_class}
\title{Agepro Model R6 class object validation}
\usage{
check_agepro_model_class(x, class_type = "agepro_inp_model")

assert_agepro_model_class(
  x,
  class_type = c("agepro_inp_model", "agepro_json_model", "agepro_model"),
  .var.name = checkmate::vname(x),
  add = NULL
)
}
\arguments{
\item{x}{Object To Check}

\item{class_type}{Type of agepro_model class type to validate. Default is \code{agepro_inp_model}}

\item{.var.name}{\code{character(1)}\cr
Name of the checked object to print in assertions. Defaults to
the heuristic implemented in
\code{\link[checkmate:vname]{checkmate::vname}}.}

\item{add}{\code{checkmate::AssertCollection}\cr
Collection to store assertion messages. See
\code{\link[checkmate:AssertCollection]{checkmate::AssertCollection}}}
}
\description{
Wrapper of checkmate::checkR6 function, to check the input object is a
"agepro_model" type R6class that includes the public fields as defined
in agepro_model.
}
