% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{check_model_bootstrap}
\alias{check_model_bootstrap}
\alias{assert_model_bootstrap}
\title{AGEPRO Model Bootstrap file path valdation}
\usage{
check_model_bootstrap(model)

assert_model_bootstrap(x, .var.name = checkmate::vname(x), add = NULL)
}
\arguments{
\item{model}{AGEPRO model class for AGEPRO Input File format (\code{agepro_inp_model})}

\item{x}{object to check}

\item{.var.name}{\code{character(1)}\cr
Name of the checked object to print in assertions. Defaults to
the heuristic implemented in
\code{\link[checkmate:vname]{checkmate::vname}}.}

\item{add}{\code{checkmate::AssertCollection}\cr
Collection to store assertion messages. See
\code{\link[checkmate:AssertCollection]{checkmate::AssertCollection}}}
}
\description{
Helper function to check bootstrap file locations. This is typically used to
check the bootstrap file path prior AGEPRO calcuation engine launch.
}
\details{
For portabilty and AGEPRO model reproducabilty, bootstrap file path strings
may be relative to the location as the input file.
}
