% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochastic.R
\name{discard_fraction}
\alias{discard_fraction}
\title{Discard fraction of numbers at age}
\description{
Class Structure for discard faction at age AGEPRO Keyword parameter.
AGEPRO model must indicate \emph{discards are present}, enabled via general
options \link[=general_params]{discards field}.
}
\section{Super class}{
\code{\link[ageproR:stochastic]{ageproR::stochastic}} -> \code{discard_fraction}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-discard_fraction-new}{\code{discard_fraction$new()}}
\item \href{#method-discard_fraction-clone}{\code{discard_fraction$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="ageproR" data-topic="stochastic" data-id="create_stochastic_table"><a href='../../ageproR/html/stochastic.html#method-stochastic-create_stochastic_table'><code>ageproR::stochastic$create_stochastic_table()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ageproR" data-topic="stochastic" data-id="inplines_stochastic"><a href='../../ageproR/html/stochastic.html#method-stochastic-inplines_stochastic'><code>ageproR::stochastic$inplines_stochastic()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ageproR" data-topic="stochastic" data-id="print"><a href='../../ageproR/html/stochastic.html#method-stochastic-print'><code>ageproR::stochastic$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ageproR" data-topic="stochastic" data-id="read_inp_lines"><a href='../../ageproR/html/stochastic.html#method-stochastic-read_inp_lines'><code>ageproR::stochastic$read_inp_lines()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ageproR" data-topic="stochastic" data-id="read_inplines_cv_table"><a href='../../ageproR/html/stochastic.html#method-stochastic-read_inplines_cv_table'><code>ageproR::stochastic$read_inplines_cv_table()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ageproR" data-topic="stochastic" data-id="read_inplines_stochastic_tables"><a href='../../ageproR/html/stochastic.html#method-stochastic-read_inplines_stochastic_tables'><code>ageproR::stochastic$read_inplines_stochastic_tables()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ageproR" data-topic="stochastic" data-id="setup_stochastic_tables"><a href='../../ageproR/html/stochastic.html#method-stochastic-setup_stochastic_tables'><code>ageproR::stochastic$setup_stochastic_tables()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-discard_fraction-new"></a>}}
\if{latex}{\out{\hypertarget{method-discard_fraction-new}{}}}
\subsection{Method \code{new()}}{
Initializes Class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{discard_fraction$new(
  proj_years,
  num_ages,
  num_fleets,
  input_option = 0,
  time_varying = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{proj_years}}{\link[=projection_years]{Projection years}:
Input can be Sequence of years in from first to last year of
projection or the number of years in the time projection.}

\item{\code{num_ages}}{Number of Age classes}

\item{\code{num_fleets}}{Number of Fleets.}

\item{\code{input_option}}{Option to indicate this parameter will be:
\itemize{
\item \code{0} By default, done interactively via interface.
\item \code{1} Imported via location of an existing data file.
}}

\item{\code{time_varying}}{Logical flag that enables the stochastic parameter
to use as a time-varying array if TRUE (or 1). Otherwise, FALSE the
vector will cover "all years" of the projection. Default is TRUE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-discard_fraction-clone"></a>}}
\if{latex}{\out{\hypertarget{method-discard_fraction-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{discard_fraction$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
