% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruit_model.R
\name{empirical_cdf_model}
\alias{empirical_cdf_model}
\title{Empirical CDF of Recruitment (Model #14)}
\description{
Empirical CDF of Recruitment (Model #14)

Empirical CDF of Recruitment (Model #14)
}
\section{Super classes}{
\code{\link[ageproR:recruit_model]{ageproR::recruit_model}} -> \code{\link[ageproR:empirical_recruit]{ageproR::empirical_recruit}} -> \code{empirical_cdf_model}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-empirical_cdf_model-new}{\code{empirical_cdf_model$new()}}
\item \href{#method-empirical_cdf_model-clone}{\code{empirical_cdf_model$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="ageproR" data-topic="empirical_recruit" data-id="inplines_recruit_data"><a href='../../ageproR/html/empirical_recruit.html#method-empirical_recruit-inplines_recruit_data'><code>ageproR::empirical_recruit$inplines_recruit_data()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ageproR" data-topic="empirical_recruit" data-id="new_obs_table"><a href='../../ageproR/html/empirical_recruit.html#method-empirical_recruit-new_obs_table'><code>ageproR::empirical_recruit$new_obs_table()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ageproR" data-topic="empirical_recruit" data-id="print"><a href='../../ageproR/html/empirical_recruit.html#method-empirical_recruit-print'><code>ageproR::empirical_recruit$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ageproR" data-topic="empirical_recruit" data-id="print_json"><a href='../../ageproR/html/empirical_recruit.html#method-empirical_recruit-print_json'><code>ageproR::empirical_recruit$print_json()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ageproR" data-topic="empirical_recruit" data-id="read_inp_lines"><a href='../../ageproR/html/empirical_recruit.html#method-empirical_recruit-read_inp_lines'><code>ageproR::empirical_recruit$read_inp_lines()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-empirical_cdf_model-new"></a>}}
\if{latex}{\out{\hypertarget{method-empirical_cdf_model-new}{}}}
\subsection{Method \code{new()}}{
Initialize the Empirical CDF Model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{empirical_cdf_model$new(num_observations = 2)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{num_observations}}{\cr
Number of Empirical Observation Records}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-empirical_cdf_model-clone"></a>}}
\if{latex}{\out{\hypertarget{method-empirical_cdf_model-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{empirical_cdf_model$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
