% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruit_model.R
\name{empirical_distribution_model}
\alias{empirical_distribution_model}
\title{Empirical Recruitment Distribution (Model #3)}
\description{
Empirical Recruitment Distribution (Model #3)

Empirical Recruitment Distribution (Model #3)
}
\section{Super classes}{
\code{\link[AGEPROjson:recruit_model]{AGEPROjson::recruit_model}} -> \code{\link[AGEPROjson:empirical_recruit]{AGEPROjson::empirical_recruit}} -> \code{empirical_distribution_model}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-empirical_distribution_model-new}{\code{empirical_distribution_model$new()}}
\item \href{#method-empirical_distribution_model-clone}{\code{empirical_distribution_model$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="AGEPROjson" data-topic="empirical_recruit" data-id="new_obs_table"><a href='../../AGEPROjson/html/empirical_recruit.html#method-empirical_recruit-new_obs_table'><code>AGEPROjson::empirical_recruit$new_obs_table()</code></a></span></li>
<li><span class="pkg-link" data-pkg="AGEPROjson" data-topic="empirical_recruit" data-id="print"><a href='../../AGEPROjson/html/empirical_recruit.html#method-empirical_recruit-print'><code>AGEPROjson::empirical_recruit$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="AGEPROjson" data-topic="empirical_recruit" data-id="print_json"><a href='../../AGEPROjson/html/empirical_recruit.html#method-empirical_recruit-print_json'><code>AGEPROjson::empirical_recruit$print_json()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-empirical_distribution_model-new"></a>}}
\if{latex}{\out{\hypertarget{method-empirical_distribution_model-new}{}}}
\subsection{Method \code{new()}}{
Initialize the Empirical Recruitment Distribution Model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{empirical_distribution_model$new(seq_years)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{seq_years}}{\cr
Array representing the Projection Time Horizon}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-empirical_distribution_model-clone"></a>}}
\if{latex}{\out{\hypertarget{method-empirical_distribution_model-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{empirical_distribution_model$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
