% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruit_model.R
\name{empirical_recruit}
\alias{empirical_recruit}
\title{Empirical Recruitment Model Data}
\description{
Recruitment Model Description
}
\section{Super class}{
\code{\link[ageproj:recruit_model]{ageproj::recruit_model}} -> \code{empirical_recruit}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{rec_array}}{Recruitment Inupt Array (data)}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{with_ssb}}{with ssb}

\item{\code{low_bound}}{Lowest significant number bound}

\item{\code{recruit_data}}{gets JSON-ready Recruit Model Data}

\item{\code{observed_points}}{Gets/Sets the number of observations used of the model projection}

\item{\code{observed_year_array}}{Gets/Sets the observed years sequence used in the model projection}

\item{\code{super_}}{Binds the super class with the empirical_recruit child classes}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-empirical_recruit-new}{\code{empirical_recruit$new()}}
\item \href{#method-empirical_recruit-new_obs_table}{\code{empirical_recruit$new_obs_table()}}
\item \href{#method-empirical_recruit-print}{\code{empirical_recruit$print()}}
\item \href{#method-empirical_recruit-print_json}{\code{empirical_recruit$print_json()}}
\item \href{#method-empirical_recruit-clone}{\code{empirical_recruit$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-empirical_recruit-new"></a>}}
\if{latex}{\out{\hypertarget{method-empirical_recruit-new}{}}}
\subsection{Method \code{new()}}{
Creates an Empirical Recruit instance
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{empirical_recruit$new(rec_points, with_ssb = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rec_points}}{Number of Recruitment Observations}

\item{\code{with_ssb}}{Empirical Recruitment includes Spawning
Stock Biomass (SSB)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-empirical_recruit-new_obs_table"></a>}}
\if{latex}{\out{\hypertarget{method-empirical_recruit-new_obs_table}{}}}
\subsection{Method \code{new_obs_table()}}{
Create Obs table
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{empirical_recruit$new_obs_table()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-empirical_recruit-print"></a>}}
\if{latex}{\out{\hypertarget{method-empirical_recruit-print}{}}}
\subsection{Method \code{print()}}{
Prints out Recruitment Model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{empirical_recruit$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{further arguments passed to or from other methods}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-empirical_recruit-print_json"></a>}}
\if{latex}{\out{\hypertarget{method-empirical_recruit-print_json}{}}}
\subsection{Method \code{print_json()}}{
Returns Recruit data as JSON
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{empirical_recruit$print_json()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-empirical_recruit-clone"></a>}}
\if{latex}{\out{\hypertarget{method-empirical_recruit-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{empirical_recruit$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
