% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecruitModel.R
\name{EmpiricalRecruitModel}
\alias{EmpiricalRecruitModel}
\title{Empirical Recritment Model Data}
\description{
Recruitment Model Description
}
\section{Super class}{
\code{\link[AGEPROjson:RecruitModel]{AGEPROjson::RecruitModel}} -> \code{EmpiricalRecruitModel}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{rec_points}}{num obs}

\item{\code{with_ssb}}{with ssb}

\item{\code{low_bound}}{Lowest significant number bound}

\item{\code{rec_array}}{Recruitment Inupt Array (data)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-EmpiricalRecruitModel-new}{\code{EmpiricalRecruitModel$new()}}
\item \href{#method-EmpiricalRecruitModel-new_obs_table}{\code{EmpiricalRecruitModel$new_obs_table()}}
\item \href{#method-EmpiricalRecruitModel-print_json}{\code{EmpiricalRecruitModel$print_json()}}
\item \href{#method-EmpiricalRecruitModel-clone}{\code{EmpiricalRecruitModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EmpiricalRecruitModel-new"></a>}}
\if{latex}{\out{\hypertarget{method-EmpiricalRecruitModel-new}{}}}
\subsection{Method \code{new()}}{
Creates an Empirical Recruit instance
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EmpiricalRecruitModel$new(model_num, rec_points, with_ssb = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model_num}}{\cr
AGEPRO Recruitment Model Number}

\item{\code{rec_points}}{Number of Recruitmebt Observations}

\item{\code{with_ssb}}{Empirical Recruitment includes Spawning
Stock Biomass (SSB)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EmpiricalRecruitModel-new_obs_table"></a>}}
\if{latex}{\out{\hypertarget{method-EmpiricalRecruitModel-new_obs_table}{}}}
\subsection{Method \code{new_obs_table()}}{
Create Obs table
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EmpiricalRecruitModel$new_obs_table()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EmpiricalRecruitModel-print_json"></a>}}
\if{latex}{\out{\hypertarget{method-EmpiricalRecruitModel-print_json}{}}}
\subsection{Method \code{print_json()}}{
Returns Recuit data as JSON
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EmpiricalRecruitModel$print_json()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EmpiricalRecruitModel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-EmpiricalRecruitModel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EmpiricalRecruitModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
