% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{file_dialog}
\alias{file_dialog}
\title{Choose a file to open or save a file interactively.}
\usage{
file_dialog(
  type = c("open", "save"),
  ext = c("All Files", ".*"),
  force_tcltk = FALSE
)
}
\arguments{
\item{type}{Character option to request a "Open" or "Save file Dialog}

\item{ext}{Two item vector describing the filename description and filename extension respecively. By default it is assigned to \code{c("All Files", "*")}}

\item{force_tcltk}{Logical}
}
\value{
A Character vector representing the filepath of the chosen path from the file dialog.
}
\description{
This function allows the user to open a File Dialog interface to interatively choose to load or save a file.
Completion of this function will only return the filepath; it will not load nor save files by itself. TODO: RStudio/tcltk/vscode
}
\keyword{internal}
