% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochastic.R
\name{fishery_selectivity}
\alias{fishery_selectivity}
\title{Fishery Selectivity at age by fleet}
\description{
Class Structure for Fishery Selectivity at age by fleet AGEPRO Keyword
parameter.
}
\section{Super class}{
\code{\link[ageproR:stochastic]{ageproR::stochastic}} -> \code{fishery_selectivity}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-fishery_selectivity-initalize}{\code{fishery_selectivity$initalize()}}
\item \href{#method-fishery_selectivity-clone}{\code{fishery_selectivity$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="ageproR" data-topic="stochastic" data-id="create_stochastic_table"><a href='../../ageproR/html/stochastic.html#method-stochastic-create_stochastic_table'><code>ageproR::stochastic$create_stochastic_table()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ageproR" data-topic="stochastic" data-id="initialize"><a href='../../ageproR/html/stochastic.html#method-stochastic-initialize'><code>ageproR::stochastic$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ageproR" data-topic="stochastic" data-id="inplines_stochastic"><a href='../../ageproR/html/stochastic.html#method-stochastic-inplines_stochastic'><code>ageproR::stochastic$inplines_stochastic()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ageproR" data-topic="stochastic" data-id="print"><a href='../../ageproR/html/stochastic.html#method-stochastic-print'><code>ageproR::stochastic$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ageproR" data-topic="stochastic" data-id="read_inp_lines"><a href='../../ageproR/html/stochastic.html#method-stochastic-read_inp_lines'><code>ageproR::stochastic$read_inp_lines()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ageproR" data-topic="stochastic" data-id="read_inplines_stochastic_tables"><a href='../../ageproR/html/stochastic.html#method-stochastic-read_inplines_stochastic_tables'><code>ageproR::stochastic$read_inplines_stochastic_tables()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ageproR" data-topic="stochastic" data-id="setup_stochastic_tables"><a href='../../ageproR/html/stochastic.html#method-stochastic-setup_stochastic_tables'><code>ageproR::stochastic$setup_stochastic_tables()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fishery_selectivity-initalize"></a>}}
\if{latex}{\out{\hypertarget{method-fishery_selectivity-initalize}{}}}
\subsection{Method \code{initalize()}}{
Initalizes new instance
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fishery_selectivity$initalize(
  proj_years,
  num_ages,
  num_fleets,
  input_option,
  time_varying = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{proj_years}}{\link[=projection_years]{Projection years}:
Input can be Sequence of years in from first to last year of
projection or the number of years in the time projection.}

\item{\code{num_ages}}{Number of Age classes}

\item{\code{num_fleets}}{Number of Fleets. Default is 1}

\item{\code{input_option}}{Option to indicate this parameter will be:
\itemize{
\item \code{0} By default, done interactively via interface.
\item \code{1} Imported via location of an existing data file.
}}

\item{\code{time_varying}}{Logical flag that enables the stochastic parameter
to use as a time-varying array if TRUE (or 1). Otherwise, FALSE the
vector will cover "all years" of the projection. Default is TRUE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fishery_selectivity-clone"></a>}}
\if{latex}{\out{\hypertarget{method-fishery_selectivity-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fishery_selectivity$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
