% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_agepro_model.R
\name{import_agepro_model}
\alias{import_agepro_model}
\title{Import data from a AGEPRO Model Input File to create a new agepro_inp_model}
\usage{
import_agepro_model(file, ...)
}
\arguments{
\item{file}{AGEPRO input file name path}

\item{...}{further arguments passed to or from other methods}
}
\value{
R6class object containing the agepro_inp_model class object
}
\description{
Function wrapper to create agepro_inp_model R6class. This function retruns a
new instance of the agepro_inp_model.
}
\examples{
\dontrun{

# Load path of ageproR's included Example1 Input File
inpfile <- file.path(find.package("ageproR"),"extdata/Example1.INP")

# Create agepro_inp_model and import data from inpfile
test <- import_agepro_model(inpfile)

}
}
