% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model.R
\name{launch_model}
\alias{launch_model}
\title{Launches agepro_model to the calcuation engine}
\usage{
launch_model(model, out_dir)
}
\arguments{
\item{model}{\link[=agepro_inp_model]{"Agepro INP File Model Class Object"}}

\item{out_dir}{Output path}
}
\description{
This function will takes the \link[=agepro_inp_model]{agepro_inp_model} class,
saves it as an AGEPRO input file, launches the AGEPRO calcuation engine with it.
}
\details{
After the calcuation engine is done, AGEPRO output file will be saved to the output
directory. Function will also store AGEPRO calcuation engine logfile if requested.

This is simlar to how AGEPRO_GUI launches agepro_model to the AGEPRO calcuation engine. The bootstrap file path written
in the AGEPRO Input File is relative to the directory that input file is saved; The bootstrap filepath value does not have
to have directory paths if the Input file and bootstrap file is saved in the same directory.

The "default directory" of the outdir will be saved at the "AGEPRO" subdirectory of the \code{R_USER} directory.
}
