% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/natural_mortality.R
\name{natural_mortality}
\alias{natural_mortality}
\title{Input information for the natural mortality rate (M) at Age}
\description{
Generalized Class Structure for Natural Mortality rate of age
AGEPRO Keyword parameter.
}
\section{Super class}{
\code{\link[ageproR:stochastic]{ageproR::stochastic}} -> \code{natural_mortality}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-natural_mortality-initalize}{\code{natural_mortality$initalize()}}
\item \href{#method-natural_mortality-clone}{\code{natural_mortality$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="ageproR" data-topic="stochastic" data-id="create_stochastic_table"><a href='../../ageproR/html/stochastic.html#method-stochastic-create_stochastic_table'><code>ageproR::stochastic$create_stochastic_table()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ageproR" data-topic="stochastic" data-id="initialize"><a href='../../ageproR/html/stochastic.html#method-stochastic-initialize'><code>ageproR::stochastic$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-natural_mortality-initalize"></a>}}
\if{latex}{\out{\hypertarget{method-natural_mortality-initalize}{}}}
\subsection{Method \code{initalize()}}{
Initializes the stochastic class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{natural_mortality$initalize(
  num_projection_years,
  num_ages,
  input_option = 0,
  time_varying = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{num_projection_years}}{Numbers of years in from first to last
year of projection.}

\item{\code{num_ages}}{Number of Age classes}

\item{\code{input_option}}{Option to indicate Natural Mortality at age will be:
\itemize{
\item \code{0} By default, done interactively via interface.
\item \code{1} Imported from a the location of an existing data file
}}

\item{\code{time_varying}}{Logical flag that enables  Natural Mortality at age
to use as a time-varying array if TRUE (or 1). Otherwise, FALSE the
vector will cover "all years" of the projection. Default is TRUE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-natural_mortality-clone"></a>}}
\if{latex}{\out{\hypertarget{method-natural_mortality-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{natural_mortality$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
