% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{open_file_dialog}
\alias{open_file_dialog}
\title{Open file dialog or interface to interactively return file path.}
\usage{
open_file_dialog(filetype)
}
\arguments{
\item{filetype}{filename extension}
}
\description{
If Rconsole is currently running in Rstudio, it will use the rstudioapi
to show the file dialog window over the IDE.
}
\keyword{internal}
