% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_options.R
\name{output_options}
\alias{output_options}
\title{AGEPRO projection output options.}
\description{
Class Structure that includes user-defined options to enable auxiliary or
options to export AGEPRO output
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{output_stock_summary}}{\link[base:logical]{Logical} flag to output stock summary information}

\item{\code{output_process_error_aux_files}}{\link[base:logical]{Logical} flag to output population and fishery processes
simulated with lognormal process error (process_error parameters) to
auxiliary output files}

\item{\code{output_data_frame}}{\link[base:logical]{Logical} flag to output AGEPRO calculation engine
projection results to R \link[base:data.frame]{data.frame}. Default is
\code{1} (or TRUE) at initialization.}

\item{\code{keyword_name}}{AGEPRO keyword parameter name}

\item{\code{inp_keyword}}{Returns AGEPRO input-file formatted Parameter}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-output_options-new}{\code{output_options$new()}}
\item \href{#method-output_options-clone}{\code{output_options$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-output_options-new"></a>}}
\if{latex}{\out{\hypertarget{method-output_options-new}{}}}
\subsection{Method \code{new()}}{
Initializes the class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{output_options$new(
  summary_report = FALSE,
  process_error_aux_files = FALSE,
  export_r_data_frame = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{summary_report}}{Logical flag to enable stock summary output.}

\item{\code{process_error_aux_files}}{Logical flag to enable output of process_error auxiliary files}

\item{\code{export_r_data_frame}}{Logical flag to enable AGEPRO ouyput to data.frame}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-output_options-clone"></a>}}
\if{latex}{\out{\hypertarget{method-output_options-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{output_options$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
