% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruit_model.R
\name{parametric_autocorrelated_error}
\alias{parametric_autocorrelated_error}
\title{Parametric Recruitment Model w/ correlated lognormal error}
\description{
Recruitment Model Description
}
\section{Super classes}{
\code{\link[ageproR:recruit_model]{ageproR::recruit_model}} -> \code{\link[ageproR:parametric_curve]{ageproR::parametric_curve}} -> \code{parametric_autocorrelated_error}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{phi}}{Sets the Autocorrelated parametric Curve Parameter, phi. Returns the
current value if no argument was passed.}

\item{\code{log_residual}}{Sets the Autocorrelated parametric Curve Parameter, log_residual.
Returns the current value if no argument was passed.}

\item{\code{json_recruit_data}}{Returns JSON-ready Recruit Model Data}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-parametric_autocorrelated_error-new}{\code{parametric_autocorrelated_error$new()}}
\item \href{#method-parametric_autocorrelated_error-print}{\code{parametric_autocorrelated_error$print()}}
\item \href{#method-parametric_autocorrelated_error-clone}{\code{parametric_autocorrelated_error$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="ageproR" data-topic="parametric_curve" data-id="inp_lines_recruit_data"><a href='../../ageproR/html/parametric_curve.html#method-parametric_curve-inp_lines_recruit_data'><code>ageproR::parametric_curve$inp_lines_recruit_data()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ageproR" data-topic="parametric_curve" data-id="read_inp_lines"><a href='../../ageproR/html/parametric_curve.html#method-parametric_curve-read_inp_lines'><code>ageproR::parametric_curve$read_inp_lines()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-parametric_autocorrelated_error-new"></a>}}
\if{latex}{\out{\hypertarget{method-parametric_autocorrelated_error-new}{}}}
\subsection{Method \code{new()}}{
Initializes the Model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{parametric_autocorrelated_error$new(
  alpha = 0,
  beta = 0,
  variance = 0,
  phi = 0,
  log_residual = 0
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{alpha}}{\cr
Stock Recruitment Parameter, alpha}

\item{\code{beta}}{\cr
Stock Recruitment Parameter, beta}

\item{\code{variance}}{\cr
variance}

\item{\code{phi}}{\cr
Autocorrelation parameters}

\item{\code{log_residual}}{\cr
Log-Scale Residual for the stock recruitment fit in the time prior to
the projection}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-parametric_autocorrelated_error-print"></a>}}
\if{latex}{\out{\hypertarget{method-parametric_autocorrelated_error-print}{}}}
\subsection{Method \code{print()}}{
Prints out Parametric Data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{parametric_autocorrelated_error$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{further arguments passed to or from other methods}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-parametric_autocorrelated_error-clone"></a>}}
\if{latex}{\out{\hypertarget{method-parametric_autocorrelated_error-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{parametric_autocorrelated_error$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
