% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruit_model.R
\name{ParametricCurveModel}
\alias{ParametricCurveModel}
\title{Parametric Recruitment Model}
\description{
Recruitment Model Description
}
\section{Super class}{
\code{\link[AGEPROjson:recruit_model]{AGEPROjson::recruit_model}} -> \code{ParametricRecruitModel}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{recruit_data}}{gets JSON-ready Recruit Model Data}

\item{\code{alpha}}{\cr
Sets the Parametric Curve Parameter, alpha. Returns the
current value if no argument was passed}

\item{\code{beta}}{\cr
Sets the Parametric Curve Parameter, beta. Returns the
current value if no argument was passed}

\item{\code{variance}}{\cr
Sets the Parametric Curve Parameter, variance. Returns the
current value if no argument was passed.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ParametricRecruitModel-new}{\code{ParametricCurveModel$new()}}
\item \href{#method-ParametricRecruitModel-print}{\code{ParametricCurveModel$print()}}
\item \href{#method-ParametricRecruitModel-clone}{\code{ParametricCurveModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ParametricRecruitModel-new"></a>}}
\if{latex}{\out{\hypertarget{method-ParametricRecruitModel-new}{}}}
\subsection{Method \code{new()}}{
Instantiate Parametric Recruitment Curve Model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParametricCurveModel$new(model_num, alpha = 0, beta = 0, variance = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model_num}}{\cr
AGEPRO Recruitment Model Number}

\item{\code{alpha}}{\cr
Stock Recruitment Parameter, alpha}

\item{\code{beta}}{\cr
Stock Recruitment Parameter, beta}

\item{\code{variance}}{\cr
variance}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ParametricRecruitModel-print"></a>}}
\if{latex}{\out{\hypertarget{method-ParametricRecruitModel-print}{}}}
\subsection{Method \code{print()}}{
Prints out Parametric Data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParametricCurveModel$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{further arguments passed to or from other methods}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ParametricRecruitModel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ParametricRecruitModel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParametricCurveModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
