% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options_percentile_summary.R
\name{percentile_summary}
\alias{percentile_summary}
\title{Percentile summary of the key results of AGEPRO projection output}
\description{
Class Structure that includes user-defined options for setting
a specific percentile for the distributions of outputs.

The logical flag \code{enable_percentile_summary} allows the user to
set values to this class field: report_percentile. The flag will also notify
\code{agepro_model} if this keyword parameter is allowed to be written to
input file.

If percentile_summary is initialized with default values, it is
presumed that this keyword parameter is not used in the agepro_model.
Therefore, \code{enable_percentile_summary} is flagged as FALSE. Valid non-
default values will set this flag to TRUE.#'
}
\details{
The percentile_summary class (or \code{PERC}) is an optional keyword
parameter used in the AGEPRO input file format.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{flag}}{A R6class containing
\link[=options_flags]{flags for optional AGEPRO options}}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{report_percentile}}{User-defined percentile for reporting the percentile of the projected
distribution of the following quantities of interest by year:
spawning stock biomass, stock biomass on January 1st, mean biomass,
combined catch biomass, landings, fishing mortality, and stock
numbers at age}

\item{\code{enable_percentile_summary}}{Read-only logical field that flags if fields can be edited. To set
the value use \code{set_enable_percentile_summary} or field}

\item{\code{json_list_object}}{Returns JSON list object of containing output_options values}

\item{\code{keyword_name}}{AGEPRO keyword parameter name}

\item{\code{inp_keyword}}{Returns AGEPRO input-file formatted Parameter}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-percentile_summary-new}{\code{percentile_summary$new()}}
\item \href{#method-percentile_summary-print}{\code{percentile_summary$print()}}
\item \href{#method-percentile_summary-set_enable_percentile_summary}{\code{percentile_summary$set_enable_percentile_summary()}}
\item \href{#method-percentile_summary-read_inp_lines}{\code{percentile_summary$read_inp_lines()}}
\item \href{#method-percentile_summary-get_inp_lines}{\code{percentile_summary$get_inp_lines()}}
\item \href{#method-percentile_summary-clone}{\code{percentile_summary$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-percentile_summary-new"></a>}}
\if{latex}{\out{\hypertarget{method-percentile_summary-new}{}}}
\subsection{Method \code{new()}}{
Initializes the class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{percentile_summary$new(perc = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{perc}}{User-defined percentile of projected distributions

When agepro_model is reinitialized, reset the value for this class's
option_flag to NULL to cleanup any values it retained previously.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-percentile_summary-print"></a>}}
\if{latex}{\out{\hypertarget{method-percentile_summary-print}{}}}
\subsection{Method \code{print()}}{
Formatted to print out output_option values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{percentile_summary$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-percentile_summary-set_enable_percentile_summary"></a>}}
\if{latex}{\out{\hypertarget{method-percentile_summary-set_enable_percentile_summary}{}}}
\subsection{Method \code{set_enable_percentile_summary()}}{
Wrapper Function to toggle enable_percentile_summary options_flag.

The percentile_summary class will not accept values until it is
enable_percentile_summary is TRUE.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{percentile_summary$set_enable_percentile_summary(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Logical value for enable_percentile_summary options_flag}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-percentile_summary-read_inp_lines"></a>}}
\if{latex}{\out{\hypertarget{method-percentile_summary-read_inp_lines}{}}}
\subsection{Method \code{read_inp_lines()}}{
Reads in the values from the keyword parameter PERC from the
AGEPRO Input file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{percentile_summary$read_inp_lines(inp_con, nline)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{inp_con}}{Open file connection to AGEPRO Input File.}

\item{\code{nline}}{\cr
Location of the current line number being read when the file connection to
the AGEPRO input file is open.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-percentile_summary-get_inp_lines"></a>}}
\if{latex}{\out{\hypertarget{method-percentile_summary-get_inp_lines}{}}}
\subsection{Method \code{get_inp_lines()}}{
Returns values from the class to the PERC AGEPRO keyword parameter
formatted as AGEPRO input file lines.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{percentile_summary$get_inp_lines(delimiter = " ")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{delimiter}}{Character string delimiter separating values of AGEPRO input file line.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-percentile_summary-clone"></a>}}
\if{latex}{\out{\hypertarget{method-percentile_summary-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{percentile_summary$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
