% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection_analyses.R
\name{pstar_projection}
\alias{pstar_projection}
\title{Projection analyses that shows the probability of exceeding overfishing
threshold of the target year}
\description{
Input information for calculating Total Allowable Catch (\eqn{TAC_{pstar}})
to produce \eqn{P*}, which is the probability of overfishing in the target
projection year
}
\keyword{projection_analyses}
\section{Super class}{
\code{\link[ageproR:projection_analyses]{ageproR::projection_analyses}} -> \code{pstar_projection}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{num_pstar_levels}}{Number of pstar values to be evaluated}

\item{\code{pstar_levels_table}}{The vector of probabilities of overfishing or PStar values to be used}

\item{\code{pstar_overfishing_f}}{Fishing mortality rate that defines the overfishing level}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-pstar_projection-new}{\code{pstar_projection$new()}}
\item \href{#method-pstar_projection-clone}{\code{pstar_projection$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pstar_projection-new"></a>}}
\if{latex}{\out{\hypertarget{method-pstar_projection-new}{}}}
\subsection{Method \code{new()}}{
Initializes class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pstar_projection$new(proj_years, num_pstar_levels = 1, pstar_f = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{proj_years}}{\link[=projection_years]{Projection years}:
Input can be Sequence of years in from first to last year of
projection or the number of years in the time projection.}

\item{\code{num_pstar_levels}}{Number of Pstar levels. Default is \code{1}}

\item{\code{pstar_f}}{Fishing mortality rate \eqn{f}. Default is \code{0.0}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pstar_projection-clone"></a>}}
\if{latex}{\out{\hypertarget{method-pstar_projection-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pstar_projection$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
