% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection_analyses.R
\name{rebuild_projection}
\alias{rebuild_projection}
\title{Input information for calculating F to rebuild spawning biomass}
\description{
Rebuilding projection analysis is focused on the calculation of the constant
total fishing mortality calculated across all fleets that will rebuild the
population, denoted as \eqn{F_{REBUILD}}.
}
\keyword{projection_analyses}
\section{Super class}{
\code{\link[ageproR:projection_analyses]{ageproR::projection_analyses}} -> \code{rebuild_projection}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{target_biomass_value}}{Rebuilding projection's target biomass value in units of thousands
of metric tons (MT)}

\item{\code{target_biomass_type}}{Index for the type of population biomass as the target:
\itemize{
\item{0}{Spawning Stock Biomass}
\item{1}{January 1st Stock Biomass}
\item{2}{Mid-Year (Mean) Biomass}
}}

\item{\code{target_percent}}{The percent frequency of achieving the target value by the target
year. The percent frequency is a value between 0 (a zero
chance of achieving target) and 100 (indicating a 100 percent chance
of achieving target).}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-rebuild_projection-new}{\code{rebuild_projection$new()}}
\item \href{#method-rebuild_projection-clone}{\code{rebuild_projection$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rebuild_projection-new"></a>}}
\if{latex}{\out{\hypertarget{method-rebuild_projection-new}{}}}
\subsection{Method \code{new()}}{
Initializes class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rebuild_projection$new(proj_years)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{proj_years}}{May be a single numeric value: the number of years in the
time projection; a vector of sequential values: Sequence of years in from
first to last year of the time projection; or an instance of
\link[=projection_years]{Projection years}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rebuild_projection-clone"></a>}}
\if{latex}{\out{\hypertarget{method-rebuild_projection-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rebuild_projection$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
