% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recruitment.R
\name{Recruitment}
\alias{Recruitment}
\title{AGEPRO Recruitment Parameter}
\description{
Recruitment is the primary stochastic element in the AGEPRO model (Brodziak,
2022). AGEPRO handles 21 Recruitment stochastic recruitment models for
projections. Please refer to the AGEPRO reference Manual for more detail on
these models.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{rec_fac}}{multiplier to convert recruitment submodel's recruitment
units to absolute numbers of fish}

\item{\code{ssb_fac}}{multiplier to convert recruitment submodel's Spawning
Biomass (SSB) to absolute spawning weight of fish in kilograms}

\item{\code{max_rec_obs}}{Recruitment submodel's maximum number of observations}

\item{\code{recruit_type}}{Recruitment Type}

\item{\code{recruit_prob}}{recruitment probabilities}

\item{\code{model_collection}}{List of recruitment models}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Recruitment-new}{\code{Recruitment$new()}}
\item \href{#method-Recruitment-get_recruit_data}{\code{Recruitment$get_recruit_data()}}
\item \href{#method-Recruitment-clone}{\code{Recruitment$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Recruitment-new"></a>}}
\if{latex}{\out{\hypertarget{method-Recruitment-new}{}}}
\subsection{Method \code{new()}}{
Initializes the Recruitment Class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Recruitment$new(model_num, seq_years)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model_num}}{\cr
AGEPRO Recruitment Model Number}

\item{\code{seq_years}}{\cr
Array representing the Projection Time Horizon}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Recruitment-get_recruit_data"></a>}}
\if{latex}{\out{\hypertarget{method-Recruitment-get_recruit_data}{}}}
\subsection{Method \code{get_recruit_data()}}{
Gets Recruitment Data


    #' set recruitment
    #'
    #' Writes the generalized AGEPRO Recruitment json data object.
    #'
    #' TODO: Recruitment models will call this function.
    #'
    #' @param model Recruitment model
    #' @param recFac Recruitment Scaling Factor
    #' @param ssbFac Spawning Stock Biomass (SSB) Scaling Factor
    #' @param maxRecObs Limits on Maximun Records Observed
    #' @param prob Recruitment Probability over each time period
    #' @param recruit Recruitment
    #'
    set_recruitment <- function (model, recFac, ssbFac, maxRecObs, prob, recruit) {



      return(list("recFac": recFac,
                  "ssbFac": ssbFac,
                  "maxRecObs": maxRecObs,
                  "type": model,
                  "prob": prob,
                  "recruit":recruit) ) #
      # "recruit": {
      #   "recFac": 1,
      #   "ssbFac": 75,
      #   "maxRecObs": 1000000,
      #   "type": 14,
      #   "prob": ["1", "1", "1", "1", "1", "1"],
      #   "submodel": {
      #     "points": 53,
      #     "recruits": [73.5939, 78.1845, 70.6004, 62.1267, 66.0886, 69.9814, 49.9445, 70.4022, 42.6731, 85.2977, 48.2887, 98.1364, 76.867, 33.8211,
      #                  7.8195, 4.3288, 2.6275, 2.7917, 4.2174, 249.227, 6.5051, 2.5329, 1.9038, 1.7011, 1.5596, 2.2002, 52.7585, 2.4754, 2.8037, 10.179,
      #                  21.2349, 8.6637, 20.0313, 11.1925, 5.0913, 4.3675, 28.9894, 51.3917, 8.7334, 35.7165, 327.489, 73.3318, 35.0047, 22.4337, 24.9481,
      #                  32.1726, 34.4703, 29.245, 81.7098, 30.5807, 25.3895, 26.28, 30.1793]
      #   }
      # },

    }
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Recruitment$get_recruit_data(model_num, seq_years)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model_num}}{\cr
AGEPRO Recruitment Model Number}

\item{\code{seq_years}}{\cr
Array representing the Projection Time Horizon}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Recruitment-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Recruitment-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Recruitment$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
