% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruitment.R
\name{recruitment}
\alias{recruitment}
\title{AGEPRO Recruitment Parameter}
\description{
Recruitment is the primary stochastic element in the AGEPRO model (Brodziak,
2022). AGEPRO handles 21 Recruitment stochastic recruitment models for
projections. Please refer to the AGEPRO reference Manual for more detail on
these models.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{recruit_model_num_list}}{Recruitment Type
#' @field recruit_probability The Recruitment Probabilities.
recruit_probability = NULL,}

\item{\code{model_collection_list}}{List of recruitment models}

\item{\code{observation_years}}{Sequence of projected years}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{max_recruit_obs}}{Recruitment submodel's maximum number of observations}

\item{\code{recruit_probability}}{The Recruitment Probabilities.}

\item{\code{num_recruit_models}}{Returns number of recruitment models}

\item{\code{recruit_scaling_factor}}{The multiplier to convert recruitment submodel's recruitment units to
absolute numbers of fish}

\item{\code{ssb_scaling_factor}}{The multiplier to convert recruitment submodel's SSB to absolute
spawning weight of fish in kilograms (kg)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-recruitment-new}{\code{recruitment$new()}}
\item \href{#method-recruitment-set_recruit_data}{\code{recruitment$set_recruit_data()}}
\item \href{#method-recruitment-set_recruit_model}{\code{recruitment$set_recruit_model()}}
\item \href{#method-recruitment-set_recruit_probability}{\code{recruitment$set_recruit_probability()}}
\item \href{#method-recruitment-print}{\code{recruitment$print()}}
\item \href{#method-recruitment-print_recruit}{\code{recruitment$print_recruit()}}
\item \href{#method-recruitment-view_recruit_data}{\code{recruitment$view_recruit_data()}}
\item \href{#method-recruitment-clone}{\code{recruitment$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-recruitment-new"></a>}}
\if{latex}{\out{\hypertarget{method-recruitment-new}{}}}
\subsection{Method \code{new()}}{
Initializes the Recruitment Class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{recruitment$new(model_num, seq_years, max_rec_obs = 10000, cat_verbose = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model_num}}{\cr
AGEPRO Recruitment Model Number}

\item{\code{seq_years}}{\cr
Array representing the Projection Time Horizon}

\item{\code{max_rec_obs}}{Max limit of recruitment observations. Default is 10000.}

\item{\code{cat_verbose}}{Flag to print out `cat` based cli messages printed on console. Default
is TRUE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-recruitment-set_recruit_data"></a>}}
\if{latex}{\out{\hypertarget{method-recruitment-set_recruit_data}{}}}
\subsection{Method \code{set_recruit_data()}}{
Creates Recruitment Model Data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{recruitment$set_recruit_data(model_num, seq_years)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model_num}}{\cr
AGEPRO Recruitment Model Number}

\item{\code{seq_years}}{\cr
Array representing the Projection Time Horizon}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-recruitment-set_recruit_model"></a>}}
\if{latex}{\out{\hypertarget{method-recruitment-set_recruit_model}{}}}
\subsection{Method \code{set_recruit_model()}}{
Initializes RecruitModel Data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{recruitment$set_recruit_model(model_num, seq_years)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model_num}}{\cr
AGEPRO Recruitment Model Number}

\item{\code{seq_years}}{\cr
Array representing the Projection Time Horizon}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-recruitment-set_recruit_probability"></a>}}
\if{latex}{\out{\hypertarget{method-recruitment-set_recruit_probability}{}}}
\subsection{Method \code{set_recruit_probability()}}{
Sets the recruitment probability
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{recruitment$set_recruit_probability(j, year, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{j}}{Index of the recruitment collection list}

\item{\code{year}}{index of the Time series}

\item{\code{value}}{Recruitment Probability}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-recruitment-print"></a>}}
\if{latex}{\out{\hypertarget{method-recruitment-print}{}}}
\subsection{Method \code{print()}}{
Prints out Recruitment
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{recruitment$print(cat_verbose = TRUE, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cat_verbose}}{Flag to allow `cat` based cli messages printed on
console. Default is TRUE}

\item{\code{...}}{further arguments passed to or from other methods}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-recruitment-print_recruit"></a>}}
\if{latex}{\out{\hypertarget{method-recruitment-print_recruit}{}}}
\subsection{Method \code{print_recruit()}}{
Prints out Recruitment object data to console, with an option to display
recruit object data to JSON format.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{recruitment$print_recruit(print_json = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{print_json}}{Option to print recruitment object as written in JSON
format into console}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-recruitment-view_recruit_data"></a>}}
\if{latex}{\out{\hypertarget{method-recruitment-view_recruit_data}{}}}
\subsection{Method \code{view_recruit_data()}}{
Helper Function To View Recruitment Model Collection Data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{recruitment$view_recruit_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-recruitment-clone"></a>}}
\if{latex}{\out{\hypertarget{method-recruitment-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{recruitment$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
