% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruitment.R
\name{recruitment}
\alias{recruitment}
\title{AGEPRO Recruitment Parameter}
\description{
Recruitment is the primary stochastic element in the AGEPRO model (Brodziak,
2022). AGEPRO handles 21 Recruitment stochastic recruitment models for
projections. Please refer to the AGEPRO reference Manual for more detail on
these models.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{recruit_scaling_factor}}{The multiplier to convert recruitment submodel's recruitment units to
absolute numbers of fish}

\item{\code{ssb_scaling_factor}}{The multiplier to convert recruitment submodel's SSB to absolute
spawning weight of fish in kilograms (kg)}

\item{\code{max_recruit_obs}}{Recruitment submodel's maximum number of observations}

\item{\code{observation_years}}{Sequence of projected years}

\item{\code{recruit_model_num_list}}{Helper Function To View Recruitment Model Collection Data}

\item{\code{num_recruit_models}}{Returns number of recruitment models}

\item{\code{recruit_probability}}{The Recruitment Probabilities.}

\item{\code{model_collection_list}}{List of recruitment models. Use this field
to access a specific recruitment models field.}

\item{\code{recruit_data}}{List containing data for each recruitment model in the recruitment
model collection list. Use this field to access a specific recruitment models field.}

\item{\code{json_list_recruit}}{List of RECRUIT keyword fields values, exportable to JSON.}

\item{\code{keyword_name}}{AGEPRO keyword parameter name}

\item{\code{inp_keyword}}{Returns AGEPRO input-file formatted Parameter}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-recruitment-new}{\code{recruitment$new()}}
\item \href{#method-recruitment-set_recruit_model}{\code{recruitment$set_recruit_model()}}
\item \href{#method-recruitment-print}{\code{recruitment$print()}}
\item \href{#method-recruitment-read_inp_lines}{\code{recruitment$read_inp_lines()}}
\item \href{#method-recruitment-inplines_recruit}{\code{recruitment$inplines_recruit()}}
\item \href{#method-recruitment-clone}{\code{recruitment$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-recruitment-new"></a>}}
\if{latex}{\out{\hypertarget{method-recruitment-new}{}}}
\subsection{Method \code{new()}}{
Initializes the Recruitment Class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{recruitment$new(
  model_num,
  seq_years,
  num_recruit_models = 1,
  max_recruit_obs = 10000,
  enable_cat_print = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model_num}}{\cr
AGEPRO Recruitment Model Number}

\item{\code{model_num}}{\cr
AGEPRO Recruitment Model Number}

\item{\code{seq_years}}{\cr
Array representing the Projection Time Horizon}

\item{\code{num_recruit_models}}{Number of Recruitment Models in AGEPRO model. Default is 1.}

\item{\code{max_recruit_obs}}{Max limit of recruitment observations. Default is 10000.}

\item{\code{enable_cat_print}}{Flag to print out \code{cat} based cli messages printed on console. Default
is TRUE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-recruitment-set_recruit_model"></a>}}
\if{latex}{\out{\hypertarget{method-recruitment-set_recruit_model}{}}}
\subsection{Method \code{set_recruit_model()}}{
Returns a recruitment model data class with default values. The type of
recruitment model is determined by the \code{model_num} parameter.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{recruitment$set_recruit_model(model_num)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model_num}}{\cr
AGEPRO Recruitment Model Number}

\item{\code{model_num}}{\cr
AGEPRO Recruitment Model Number}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-recruitment-print"></a>}}
\if{latex}{\out{\hypertarget{method-recruitment-print}{}}}
\subsection{Method \code{print()}}{
Prints out Recruitment
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{recruitment$print(enable_cat_print = TRUE, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{enable_cat_print}}{Flag to allow \code{cat} based cli messages printed on
console. Default is TRUE}

\item{\code{...}}{further arguments passed to or from other methods}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-recruitment-read_inp_lines"></a>}}
\if{latex}{\out{\hypertarget{method-recruitment-read_inp_lines}{}}}
\subsection{Method \code{read_inp_lines()}}{
Reads in Recruitment AGEPRO parameters from AGEPRO INP Input File
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{recruitment$read_inp_lines(
  inp_con,
  nline,
  seq_years = 1,
  num_recruit_models = 1
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{inp_con}}{Open file connection to AGEPRO Input File.}

\item{\code{nline}}{\cr
Location of the current line number being read when the file connection to
the AGEPRO input file is open.}

\item{\code{seq_years}}{\cr
Array representing the Projection Time Horizon}

\item{\code{num_recruit_models}}{Number of Recruitment Models. Default is 1}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-recruitment-inplines_recruit"></a>}}
\if{latex}{\out{\hypertarget{method-recruitment-inplines_recruit}{}}}
\subsection{Method \code{inplines_recruit()}}{
Returns the values for the RECRUIT keyword parameter formatted
to the AGEPRO input file format.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{recruitment$inplines_recruit(delimiter = " ")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{delimiter}}{Character string delimiter separating values of AGEPRO input file line.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-recruitment-clone"></a>}}
\if{latex}{\out{\hypertarget{method-recruitment-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{recruitment$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
