% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recruitment.R
\name{Recruitment}
\alias{Recruitment}
\title{AGEPRO Recruitment Parameter}
\description{
Recruitment is the primary stochastic element in the AGEPRO model (Brodziak,
2022). AGEPRO handles 21 Recruitment stochastic recruitment models for
projections. Please refer to the AGEPRO reference Manual for more detail on
these models.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{rec_fac}}{multiplier to convert recruitment submodel's recruitment
units to absolute numbers of fish}

\item{\code{ssb_fac}}{multiplier to convert recruitment submodel's Spawning
Biomass (SSB) to absolute spawning weight of fish in kilograms}

\item{\code{max_rec_obs}}{Recruitment submodel's maximum number of observations}

\item{\code{rec_model_num}}{Recruitment Type}

\item{\code{rec_prob}}{recruitment probabilities}

\item{\code{model_collection_list}}{List of recruitment models}

\item{\code{seq_yrs}}{Sequence of projected years}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Recruitment-new}{\code{Recruitment$new()}}
\item \href{#method-Recruitment-set_recruit_data}{\code{Recruitment$set_recruit_data()}}
\item \href{#method-Recruitment-get_recruit_data}{\code{Recruitment$get_recruit_data()}}
\item \href{#method-Recruitment-get_json}{\code{Recruitment$get_json()}}
\item \href{#method-Recruitment-view_recruit_data}{\code{Recruitment$view_recruit_data()}}
\item \href{#method-Recruitment-clone}{\code{Recruitment$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Recruitment-new"></a>}}
\if{latex}{\out{\hypertarget{method-Recruitment-new}{}}}
\subsection{Method \code{new()}}{
Initializes the Recruitment Class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Recruitment$new(model_num, seq_years)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model_num}}{\cr
AGEPRO Recruitment Model Number}

\item{\code{seq_years}}{\cr
Array representing the Projection Time Horizon}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Recruitment-set_recruit_data"></a>}}
\if{latex}{\out{\hypertarget{method-Recruitment-set_recruit_data}{}}}
\subsection{Method \code{set_recruit_data()}}{
Creates Recruitment Model Data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Recruitment$set_recruit_data(model_num, seq_years)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model_num}}{\cr
AGEPRO Recruitment Model Number}

\item{\code{seq_years}}{\cr
Array representing the Projection Time Horizon}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Recruitment-get_recruit_data"></a>}}
\if{latex}{\out{\hypertarget{method-Recruitment-get_recruit_data}{}}}
\subsection{Method \code{get_recruit_data()}}{
Gets Recruitment Data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Recruitment$get_recruit_data(model_num, seq_years)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model_num}}{\cr
AGEPRO Recruitment Model Number}

\item{\code{seq_years}}{\cr
Array representing the Projection Time Horizon}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Recruitment-get_json"></a>}}
\if{latex}{\out{\hypertarget{method-Recruitment-get_json}{}}}
\subsection{Method \code{get_json()}}{
Prints out Recruitment object data as JSON
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Recruitment$get_json(print = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{print}}{Option to print recruitment object as written in JSON
format into console}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Recruitment-view_recruit_data"></a>}}
\if{latex}{\out{\hypertarget{method-Recruitment-view_recruit_data}{}}}
\subsection{Method \code{view_recruit_data()}}{
Helper Function To View Recruitment Model Collection Data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Recruitment$view_recruit_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Recruitment-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Recruitment-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Recruitment$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
