% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_points.R
\name{reference_points}
\alias{reference_points}
\title{Values to output optional agepro model's reference point threshold report.}
\description{
Class Structure that defines biomass thresholds for spawaning biomass
\eqn{(B_{S,\ THRESHOLD})}, mean biomass \eqn{(\bar B_{S,\ THRESHOLD})},
and total stock biomass\eqn{(B_{T,\ THRESHOLD})}, including the threshold
rate for fishing mortality \eqn{(F_{THRESHOLD})}.

The logical flag \code{enable_reference_points} allows the user to
set values to this class fields. The flag will also notify \code{agepro_model}
if this keyword parameter is allowed to be written to input file.

If this class is initialized with default values, it is presumed that this
keyword parameter is not used in the agepro_model. Therefore,
\code{enable_reference_points} is flagged as FALSE. Valid non-default values will
set this flag to TRUE.
}
\details{
The reference_points class (\code{BOUNDS}) is recognized as a keyword
parameter used in the AGEPRO input file format, but it is optional.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{flag}}{R6class containing option_flags}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{ssb_threshold}}{Threshold of spawning biomass}

\item{\code{stock_biomass_threshold}}{Threshold of total stock biomass}

\item{\code{mean_biomass_threshold}}{Threshold for mean biomass}

\item{\code{fishing_mortality_threshold}}{Threshold for the fishing mortality rate for annual total fishing
mortality calculated across all fleets.}

\item{\code{enable_reference_points}}{Logical field that flags if fields can be edited. This class will not
accept new values to its fields or allow it to be exported to input file
until this option flag is TRUE.}

\item{\code{keyword_name}}{AGEPRO keyword parameter name}

\item{\code{inp_keyword}}{Returns AGEPRO input-file formatted Parameter}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-reference_points-initalize}{\code{reference_points$initalize()}}
\item \href{#method-reference_points-clone}{\code{reference_points$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reference_points-initalize"></a>}}
\if{latex}{\out{\hypertarget{method-reference_points-initalize}{}}}
\subsection{Method \code{initalize()}}{
Initializes the class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{reference_points$initalize(
  ssb_threshold = 0,
  stockbio_threshold = 0,
  meanbio_threshold = 0,
  fmort_threshold = 0
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ssb_threshold}}{Spawning Biomass threshold express in biomass output units (MT).}

\item{\code{stockbio_threshold}}{Stock biomass threshold expressed in biomass output units(MT).}

\item{\code{meanbio_threshold}}{Mean biomass threshold expressed in biomass output units (MT)}

\item{\code{fmort_threshold}}{Fishing mortality threshold}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reference_points-clone"></a>}}
\if{latex}{\out{\hypertarget{method-reference_points-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{reference_points$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
