% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model.R
\name{run_model}
\alias{run_model}
\title{Run the AGEPRO Cacluation Engine Binary with a AGEPRO Input File}
\usage{
run_model(
  exepath = "agepro.exe",
  agepro_args = "",
  outdir = getwd(),
  save_logfile = FALSE
)
}
\arguments{
\item{exepath}{Path of the Agepro Calcuation engine}

\item{agepro_args}{Typically used for AGEPRO Input Files.}

\item{outdir}{Output path}

\item{save_logfile}{Option to save AGEPRO Calcuation Engine ouput to logfile}
}
\description{
Wrapper function to call the Binary
}
\examples{
\dontrun{

inp_file <- system.file("extdata/Example1.INP", package = "ageproR")

# For this example, save the Agepro Calcuation Engine Binary in
# the 'agepro' subdirectory of home directory 'R_USER'
exepath <- path.expand("~/agepro/AGEPRO.EXE")
if(.Platform$OS.type == "windows") {
  exepath <- normalizePath(exepath, winslash="/")
}

ageproR::run_model(exepath, inp_file, tempdir())
}

}
