% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model.R
\name{run_model}
\alias{run_model}
\title{Run an AGEPRO Input File}
\usage{
run_model(
  exepath = "agepro.exe",
  agepro_args = "",
  outdir = getwd(),
  save_logfile = FALSE,
  fn_logfile = tempfile(format(Sys.time(), "\%Y\%m\%d_\%H\%M_"), fileext = ".txt"),
  verbose = TRUE,
  ...
)
}
\arguments{
\item{exepath}{Path of the Agepro Calcuation engine}

\item{agepro_args}{Typically used for AGEPRO Input Files.}

\item{outdir}{Output path}

\item{save_logfile}{Option to save AGEPRO Calcuation Engine ouput to logfile}

\item{fn_logfile}{Logfile Filename}

\item{verbose}{Option to show verbose output}

\item{...}{further arguments passed to or from other methods}
}
\description{
Function
}
\examples{
\dontrun{

wdir <- here::here()
inpfile <- file.path(wdir,)
ageproR::run_model(wdir,"AGEPRO.EXE",)
}
}
