% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling_factors.R
\name{scaling_factors}
\alias{scaling_factors}
\title{Scale Factors}
\description{
Input information on scaling factors for biomass, recruitment, and stock size
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{flag}}{R6class containing option_flags}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{biomass_scale}}{Output units of biomass expressed in thousand metric tons}

\item{\code{recruitment_scale}}{Output units of recruitment numbers}

\item{\code{stock_size_scale}}{Output Units of stock size numbers}

\item{\code{enable_scaling_factors}}{Logical field that flags if fields can be edited. This class will not
accept new values to its fields or allow it to be exported to input file
until this option flag is TRUE.}

\item{\code{keyword_name}}{AGEPRO keyword parameter name}

\item{\code{inp_keyword}}{Returns AGEPRO input-file formatted Parameter}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-scaling_factors-new}{\code{scaling_factors$new()}}
\item \href{#method-scaling_factors-print}{\code{scaling_factors$print()}}
\item \href{#method-scaling_factors-clone}{\code{scaling_factors$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-scaling_factors-new"></a>}}
\if{latex}{\out{\hypertarget{method-scaling_factors-new}{}}}
\subsection{Method \code{new()}}{
Initializes the class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{scaling_factors$new(scale_bio = 0, scale_recruit = 0, scale_stock_size = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{scale_bio}}{Output units of biomass express in thousand metric units}

\item{\code{scale_recruit}}{Output Units of Recruitment Numbers}

\item{\code{scale_stock_size}}{Output Units of Stock Size Numbers}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-scaling_factors-print"></a>}}
\if{latex}{\out{\hypertarget{method-scaling_factors-print}{}}}
\subsection{Method \code{print()}}{
Formatted to print out scaling_factors values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{scaling_factors$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-scaling_factors-clone"></a>}}
\if{latex}{\out{\hypertarget{method-scaling_factors-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{scaling_factors$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
