% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeInput.R
\name{set_bootstrap}
\alias{set_bootstrap}
\title{Bootstrapping}
\usage{
set_bootstrap(
  nboot = 1000,
  bootFac = 1000,
  file = here("inst", "Example1.BSN")
)
}
\arguments{
\item{nboot}{Number Of Bootstrap Operations}

\item{bootFac}{Population scaling Factor}

\item{file}{Bootstrap File}
}
\description{
AGEPRO Bootstrapping options
}
