% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeInput.R
\name{set_general}
\alias{set_general}
\title{AGEPRO general options}
\usage{
set_general(
  nFYear,
  nXYear,
  nFAge,
  nXAge,
  nSims,
  nFleet,
  nRecModel,
  seed,
  discard = FALSE
)
}
\arguments{
\item{nFYear}{First Year of prediction}

\item{nXYear}{Last Year of prediction}

\item{nFAge}{First Age Class}

\item{nXAge}{Last Age Class}

\item{nSims}{Number of Population Simulations}

\item{nFleet}{Number of Fleets}

\item{nRecModel}{Number of Recruitment Models}

\item{seed}{Psuedo-random number seed}

\item{discard}{Are Discards Are Present}
}
\description{
Sets general options for AGEPRO
}
