% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_input.R
\name{set_version}
\alias{set_version}
\title{AGEPRO version}
\usage{
set_version(
  legacy_version_string = "AGEPRO VERSION 4.0",
  version_number = "4.0.0.0"
)
}
\arguments{
\item{legacy_version_string}{Version Character String generated by legacy
versions}

\item{version_number}{AGEPRO version}
}
\description{
Sets/marks AGEPRO Calcuation Version
}
