% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochastic.R
\name{stochastic}
\alias{stochastic}
\title{Stochastic AGEPRO Keyword Parameter Structure.}
\description{
Generalized Class Structure for Stochastic AGEPRO Keyword parameters.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{input_option}}{Stochastic Input option}

\item{\code{time_varying}}{Logical flag to list stochastic data by observation
year}

\item{\code{stochastic_table}}{This is the logic for the average stochastic
AGEPRO keyword parameter's at age (and by fleet if fleets are a
factor).}

\item{\code{cv_table}}{Matrix containing the vector of of age-specific CVs for
sampling the average stochastic AGEPRO keyword parameter's at age
(and by fleet if fleets are a factor) with lognormal process error.}

\item{\code{parameter_name}}{Parameter Name}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-stochastic-new}{\code{stochastic$new()}}
\item \href{#method-stochastic-setup_stochastic_tables}{\code{stochastic$setup_stochastic_tables()}}
\item \href{#method-stochastic-create_stochastic_table}{\code{stochastic$create_stochastic_table()}}
\item \href{#method-stochastic-print}{\code{stochastic$print()}}
\item \href{#method-stochastic-read_inp_lines}{\code{stochastic$read_inp_lines()}}
\item \href{#method-stochastic-read_inplines_stochastic_tables}{\code{stochastic$read_inplines_stochastic_tables()}}
\item \href{#method-stochastic-clone}{\code{stochastic$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-stochastic-new"></a>}}
\if{latex}{\out{\hypertarget{method-stochastic-new}{}}}
\subsection{Method \code{new()}}{
Initializes the stochastic class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{stochastic$new(
  proj_years,
  num_ages,
  num_fleets = 1,
  input_option = 0,
  time_varying = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{proj_years}}{\link[=projection_years]{Projection years}:
Input can be Sequence of years in from first to last year of
projection or the number of years in the time projection.}

\item{\code{num_ages}}{Number of Age classes}

\item{\code{num_fleets}}{Number of Fleets. Default is 1}

\item{\code{input_option}}{Option to indicate stochastic parameter will be:
\itemize{
\item \code{0} By default, interactively via interface.
\item \code{1} Imported from a the location of an existing data file
}}

\item{\code{time_varying}}{Logical flag that enables the stochastic parameter to use
as a time-varying array if TRUE (or 1). Otherwise, FALSE the vector will
cover "all years" of the projection. Default is TRUE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-stochastic-setup_stochastic_tables"></a>}}
\if{latex}{\out{\hypertarget{method-stochastic-setup_stochastic_tables}{}}}
\subsection{Method \code{setup_stochastic_tables()}}{
Initialize Stochastic and CV tables
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{stochastic$setup_stochastic_tables(
  projection_years,
  num_ages,
  num_fleets = 1,
  time_varying = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{projection_years}}{\link[=projection_years]{Projection years}
value}

\item{\code{num_ages}}{Number of Ages}

\item{\code{num_fleets}}{Number of Fleets. Defaults to 1}

\item{\code{time_varying}}{Logical flag that enables the stochastic parameter to use
as a time-varying array if TRUE (or 1). Otherwise, FALSE the vector will
cover "all years" of the projection. Default is TRUE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-stochastic-create_stochastic_table"></a>}}
\if{latex}{\out{\hypertarget{method-stochastic-create_stochastic_table}{}}}
\subsection{Method \code{create_stochastic_table()}}{
Creates an stochastic table
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{stochastic$create_stochastic_table(fleet_yr_rows, ages_cols)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fleet_yr_rows}}{(Fleet-)Year Row}

\item{\code{ages_cols}}{Age Columns}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-stochastic-print"></a>}}
\if{latex}{\out{\hypertarget{method-stochastic-print}{}}}
\subsection{Method \code{print()}}{
Formatted print out Stochastic Parameter Values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{stochastic$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{further arguments passed to or from other methods}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-stochastic-read_inp_lines"></a>}}
\if{latex}{\out{\hypertarget{method-stochastic-read_inp_lines}{}}}
\subsection{Method \code{read_inp_lines()}}{
Reads in stochastic parameter's from AGEPRO Input file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{stochastic$read_inp_lines(inp_con, nline)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{inp_con}}{Open file connection to AGEPRO Input File.}

\item{\code{nline}}{\cr
Location of the current line number being read when the file connection to
the AGEPRO input file is open.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-stochastic-read_inplines_stochastic_tables"></a>}}
\if{latex}{\out{\hypertarget{method-stochastic-read_inplines_stochastic_tables}{}}}
\subsection{Method \code{read_inplines_stochastic_tables()}}{
Internal helper function to set stochastic tables from
AGEPRO input files.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{stochastic$read_inplines_stochastic_tables(inp_con, nline)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{inp_con}}{Open file connection to AGEPRO Input File.}

\item{\code{nline}}{\cr
Location of the current line number being read when the file connection to
the AGEPRO input file is open.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-stochastic-clone"></a>}}
\if{latex}{\out{\hypertarget{method-stochastic-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{stochastic$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
