% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruit_model.R
\name{two_stage_empirical_cdf}
\alias{two_stage_empirical_cdf}
\title{Two-Stage Empirical Cumulative Distribution Function of Recruitment}
\description{
Two-Stage Empirical Cumulative Distribution Function of Recruitment

Two-Stage Empirical Cumulative Distribution Function of Recruitment
}
\section{Super classes}{
\code{\link[ageproj:recruit_model]{ageproj::recruit_model}} -> \code{\link[ageproj:empirical_recruit]{ageproj::empirical_recruit}} -> \code{two_stage_empirical_cdf}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{num_low_recruits}}{Number of Low State Recruitments}

\item{\code{num_high_recruits}}{Number of high State Recruitments}

\item{\code{ssb_cutoff}}{Cutoff level of spawning Biomass}

\item{\code{low_recruits}}{Vector of Low State Recruitments per Spawning Biomass}

\item{\code{high_recruits}}{Vector of High State Recruitments per Spawning Biomass}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-two_stage_empirical_cdf-new}{\code{two_stage_empirical_cdf$new()}}
\item \href{#method-two_stage_empirical_cdf-clone}{\code{two_stage_empirical_cdf$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="ageproj" data-topic="empirical_recruit" data-id="new_obs_table"><a href='../../ageproj/html/empirical_recruit.html#method-empirical_recruit-new_obs_table'><code>ageproj::empirical_recruit$new_obs_table()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ageproj" data-topic="empirical_recruit" data-id="print"><a href='../../ageproj/html/empirical_recruit.html#method-empirical_recruit-print'><code>ageproj::empirical_recruit$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ageproj" data-topic="empirical_recruit" data-id="print_json"><a href='../../ageproj/html/empirical_recruit.html#method-empirical_recruit-print_json'><code>ageproj::empirical_recruit$print_json()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ageproj" data-topic="empirical_recruit" data-id="read_inp_lines"><a href='../../ageproj/html/empirical_recruit.html#method-empirical_recruit-read_inp_lines'><code>ageproj::empirical_recruit$read_inp_lines()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-two_stage_empirical_cdf-new"></a>}}
\if{latex}{\out{\hypertarget{method-two_stage_empirical_cdf-new}{}}}
\subsection{Method \code{new()}}{
Initialize the Empirical CDF Model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{two_stage_empirical_cdf$new(num_observations = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{num_observations}}{\cr
Number of Empirical Observation Records}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-two_stage_empirical_cdf-clone"></a>}}
\if{latex}{\out{\hypertarget{method-two_stage_empirical_cdf-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{two_stage_empirical_cdf$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
