% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruit_model.R
\name{two_stage_empirical_cdf}
\alias{two_stage_empirical_cdf}
\title{Two-Stage Empirical Cumulative Distribution Function of Recruitment
(Recruit #15)}
\description{
Two-Stage Empirical Cumulative Distribution Function of Recruitment
(Recruit #15)

Two-Stage Empirical Cumulative Distribution Function of Recruitment
(Recruit #15)
}
\section{Super classes}{
\code{\link[ageproj:recruit_model]{ageproj::recruit_model}} -> \code{\link[ageproj:empirical_recruit]{ageproj::empirical_recruit}} -> \code{ageproj::two_stage_emirical_recruit} -> \code{two_stage_empirical_cdf}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-two_stage_empirical_cdf-new}{\code{two_stage_empirical_cdf$new()}}
\item \href{#method-two_stage_empirical_cdf-print}{\code{two_stage_empirical_cdf$print()}}
\item \href{#method-two_stage_empirical_cdf-clone}{\code{two_stage_empirical_cdf$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="ageproj" data-topic="empirical_recruit" data-id="new_obs_table"><a href='../../ageproj/html/empirical_recruit.html#method-empirical_recruit-new_obs_table'><code>ageproj::empirical_recruit$new_obs_table()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ageproj" data-topic="empirical_recruit" data-id="print_json"><a href='../../ageproj/html/empirical_recruit.html#method-empirical_recruit-print_json'><code>ageproj::empirical_recruit$print_json()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ageproj" data-topic="two_stage_emirical_recruit" data-id="read_inp_lines"><a href='../../ageproj/html/two_stage_emirical_recruit.html#method-two_stage_emirical_recruit-read_inp_lines'><code>ageproj::two_stage_emirical_recruit$read_inp_lines()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-two_stage_empirical_cdf-new"></a>}}
\if{latex}{\out{\hypertarget{method-two_stage_empirical_cdf-new}{}}}
\subsection{Method \code{new()}}{
Initialize the Empirical CDF Model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{two_stage_empirical_cdf$new(low_recruits = 1, high_recruits = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{low_recruits}}{The number of low recruits per spawning stock biomass data points.}

\item{\code{high_recruits}}{The number of high recruits per spawning stock biomass data points.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-two_stage_empirical_cdf-print"></a>}}
\if{latex}{\out{\hypertarget{method-two_stage_empirical_cdf-print}{}}}
\subsection{Method \code{print()}}{
Prints out Recruitment Model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{two_stage_empirical_cdf$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{further arguments passed to or from other methods}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-two_stage_empirical_cdf-clone"></a>}}
\if{latex}{\out{\hypertarget{method-two_stage_empirical_cdf-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{two_stage_empirical_cdf$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
