% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruit_model.R
\name{two_stage_empirical_cdf}
\alias{two_stage_empirical_cdf}
\title{Two-Stage Empirical Cumulative Distribution Function of Recruitment
(Recruit #15)}
\description{
Two-Stage Empirical Cumulative Distribution Function of Recruitment
(Recruit #15)

Two-Stage Empirical Cumulative Distribution Function of Recruitment
(Recruit #15)
}
\section{Super classes}{
\code{\link[ageproR:recruit_model]{ageproR::recruit_model}} -> \code{\link[ageproR:empirical_recruit]{ageproR::empirical_recruit}} -> \code{\link[ageproR:two_stage_empirical_recruit]{ageproR::two_stage_empirical_recruit}} -> \code{two_stage_empirical_cdf}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-two_stage_empirical_cdf-new}{\code{two_stage_empirical_cdf$new()}}
\item \href{#method-two_stage_empirical_cdf-clone}{\code{two_stage_empirical_cdf$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="ageproR" data-topic="empirical_recruit" data-id="new_obs_table"><a href='../../ageproR/html/empirical_recruit.html#method-empirical_recruit-new_obs_table'><code>ageproR::empirical_recruit$new_obs_table()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ageproR" data-topic="empirical_recruit" data-id="print_json"><a href='../../ageproR/html/empirical_recruit.html#method-empirical_recruit-print_json'><code>ageproR::empirical_recruit$print_json()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ageproR" data-topic="two_stage_empirical_recruit" data-id="inplines_recruit_data"><a href='../../ageproR/html/two_stage_empirical_recruit.html#method-two_stage_empirical_recruit-inplines_recruit_data'><code>ageproR::two_stage_empirical_recruit$inplines_recruit_data()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ageproR" data-topic="two_stage_empirical_recruit" data-id="new_recruitment_matrix"><a href='../../ageproR/html/two_stage_empirical_recruit.html#method-two_stage_empirical_recruit-new_recruitment_matrix'><code>ageproR::two_stage_empirical_recruit$new_recruitment_matrix()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ageproR" data-topic="two_stage_empirical_recruit" data-id="print"><a href='../../ageproR/html/two_stage_empirical_recruit.html#method-two_stage_empirical_recruit-print'><code>ageproR::two_stage_empirical_recruit$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ageproR" data-topic="two_stage_empirical_recruit" data-id="read_inp_lines"><a href='../../ageproR/html/two_stage_empirical_recruit.html#method-two_stage_empirical_recruit-read_inp_lines'><code>ageproR::two_stage_empirical_recruit$read_inp_lines()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-two_stage_empirical_cdf-new"></a>}}
\if{latex}{\out{\hypertarget{method-two_stage_empirical_cdf-new}{}}}
\subsection{Method \code{new()}}{
Initialize the Empirical CDF Model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{two_stage_empirical_cdf$new(low_recruits = 1, high_recruits = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{low_recruits}}{\cr
The number of low recruits per spawning stock biomass data points.}

\item{\code{high_recruits}}{\cr
The number of high recruits per spawning stock biomass data points.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-two_stage_empirical_cdf-clone"></a>}}
\if{latex}{\out{\hypertarget{method-two_stage_empirical_cdf-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{two_stage_empirical_cdf$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
