% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruit_model.R
\name{two_stage_empirical_recruit}
\alias{two_stage_empirical_recruit}
\title{Two-Stage Empirical Recruitment Base}
\description{
Two-Stage Empirical Recruitment Base

Two-Stage Empirical Recruitment Base
}
\section{Super classes}{
\code{\link[ageproj:recruit_model]{ageproj::recruit_model}} -> \code{\link[ageproj:empirical_recruit]{ageproj::empirical_recruit}} -> \code{two_stage_emirical_recruit}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{num_low_recruits}}{Number of Low State Recruitments}

\item{\code{num_high_recruits}}{Number of high State Recruitments}

\item{\code{ssb_cutoff}}{Cutoff level of spawning Biomass}

\item{\code{low_recruitment}}{Vector of Low State Recruitments per Spawning Biomass}

\item{\code{high_recruitment}}{Vector of High State Recruitments per Spawning Biomass}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-two_stage_emirical_recruit-new}{\code{two_stage_empirical_recruit$new()}}
\item \href{#method-two_stage_emirical_recruit-new_recruitment_matrix}{\code{two_stage_empirical_recruit$new_recruitment_matrix()}}
\item \href{#method-two_stage_emirical_recruit-read_inp_lines}{\code{two_stage_empirical_recruit$read_inp_lines()}}
\item \href{#method-two_stage_emirical_recruit-clone}{\code{two_stage_empirical_recruit$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="ageproj" data-topic="empirical_recruit" data-id="new_obs_table"><a href='../../ageproj/html/empirical_recruit.html#method-empirical_recruit-new_obs_table'><code>ageproj::empirical_recruit$new_obs_table()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ageproj" data-topic="empirical_recruit" data-id="print"><a href='../../ageproj/html/empirical_recruit.html#method-empirical_recruit-print'><code>ageproj::empirical_recruit$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ageproj" data-topic="empirical_recruit" data-id="print_json"><a href='../../ageproj/html/empirical_recruit.html#method-empirical_recruit-print_json'><code>ageproj::empirical_recruit$print_json()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-two_stage_emirical_recruit-new"></a>}}
\if{latex}{\out{\hypertarget{method-two_stage_emirical_recruit-new}{}}}
\subsection{Method \code{new()}}{
Initialize the Empirical CDF Model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{two_stage_empirical_recruit$new(
  low_recruits = 1,
  high_recruits = 1,
  with_ssb = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{low_recruits}}{The number of low recruits per spawning stock biomass data points.}

\item{\code{high_recruits}}{The number of high recruits per spawning stock biomass data points.}

\item{\code{with_ssb}}{flag to include Spawning Stock Biomass in Observations}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-two_stage_emirical_recruit-new_recruitment_matrix"></a>}}
\if{latex}{\out{\hypertarget{method-two_stage_emirical_recruit-new_recruitment_matrix}{}}}
\subsection{Method \code{new_recruitment_matrix()}}{
Creates a recruitment matrix
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{two_stage_empirical_recruit$new_recruitment_matrix(recruit_points)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{recruit_points}}{Number of Empirical Observation Records}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-two_stage_emirical_recruit-read_inp_lines"></a>}}
\if{latex}{\out{\hypertarget{method-two_stage_emirical_recruit-read_inp_lines}{}}}
\subsection{Method \code{read_inp_lines()}}{
Reads the two State Empirical model data from AGEPRO Input file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{two_stage_empirical_recruit$read_inp_lines(inp_con, nline)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{inp_con}}{Open file connection to AGEPRO Input File.}

\item{\code{nline}}{Line Number}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-two_stage_emirical_recruit-clone"></a>}}
\if{latex}{\out{\hypertarget{method-two_stage_emirical_recruit-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{two_stage_empirical_recruit$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
