% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruit_model.R
\name{two_stage_empirical_ssb}
\alias{two_stage_empirical_ssb}
\title{Two-Stage Empirical Recruits Per Spawning Biomass Distribution (Model #4)}
\description{
Two-Stage Empirical Recruits Per Spawning Biomass Distribution (Model #4)

Two-Stage Empirical Recruits Per Spawning Biomass Distribution (Model #4)
}
\section{Super classes}{
\code{\link[ageproj:recruit_model]{ageproj::recruit_model}} -> \code{\link[ageproj:empirical_recruit]{ageproj::empirical_recruit}} -> \code{\link[ageproj:two_stage_empirical_recruit]{ageproj::two_stage_empirical_recruit}} -> \code{two_stage_empirical_ssb}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-two_stage_empirical_ssb-new}{\code{two_stage_empirical_ssb$new()}}
\item \href{#method-two_stage_empirical_ssb-clone}{\code{two_stage_empirical_ssb$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="ageproj" data-topic="empirical_recruit" data-id="new_obs_table"><a href='../../ageproj/html/empirical_recruit.html#method-empirical_recruit-new_obs_table'><code>ageproj::empirical_recruit$new_obs_table()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ageproj" data-topic="empirical_recruit" data-id="print_json"><a href='../../ageproj/html/empirical_recruit.html#method-empirical_recruit-print_json'><code>ageproj::empirical_recruit$print_json()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ageproj" data-topic="two_stage_empirical_recruit" data-id="new_recruitment_matrix"><a href='../../ageproj/html/two_stage_empirical_recruit.html#method-two_stage_empirical_recruit-new_recruitment_matrix'><code>ageproj::two_stage_empirical_recruit$new_recruitment_matrix()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ageproj" data-topic="two_stage_empirical_recruit" data-id="print"><a href='../../ageproj/html/two_stage_empirical_recruit.html#method-two_stage_empirical_recruit-print'><code>ageproj::two_stage_empirical_recruit$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ageproj" data-topic="two_stage_empirical_recruit" data-id="read_inp_lines"><a href='../../ageproj/html/two_stage_empirical_recruit.html#method-two_stage_empirical_recruit-read_inp_lines'><code>ageproj::two_stage_empirical_recruit$read_inp_lines()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-two_stage_empirical_ssb-new"></a>}}
\if{latex}{\out{\hypertarget{method-two_stage_empirical_ssb-new}{}}}
\subsection{Method \code{new()}}{
Initialize the Empirical CDF Model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{two_stage_empirical_ssb$new(low_recruits = 1, high_recruits = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{low_recruits}}{\cr
The number of low recruits per spawning stock biomass data points.}

\item{\code{high_recruits}}{\cr
The number of high recruits per spawning stock biomass data points.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-two_stage_empirical_ssb-clone"></a>}}
\if{latex}{\out{\hypertarget{method-two_stage_empirical_ssb-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{two_stage_empirical_ssb$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
