% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model.R
\name{validate_calc_engine_binary}
\alias{validate_calc_engine_binary}
\title{Validates file path for AGEPRO Calcuation Engine Binary}
\usage{
validate_calc_engine_binary(
  exepath = file.path(getwd(), "agepro.exe"),
  verbose = FALSE
)
}
\arguments{
\item{exepath}{AGEPRO Calcuuation Path Binary}

\item{verbose}{Logical option to output to RConsole.}
}
\description{
Validates file path for AGEPRO Calcuation Engine Binary
}
