% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dialog.R
\name{validate_filetype}
\alias{validate_filetype}
\title{Checks the validity of filetype key-value pair.}
\usage{
validate_filetype(filetype)
}
\arguments{
\item{filetype}{filename extension.}
}
\description{
Checks the filetype as a 2 length vector without missing values. If filetype
as not passed in the parameter, it will return the default "{All files} {*}"
string vector.
}
\details{
The filetype key-value pair is defined as \emph{fileTypeName} \emph{extension}. This
is used to \href{https://www.tcl.tk/man/tcl8.0/TkCmd/getOpenFile.html#M11}{specifying flie patterns} for
Tcl/TK file modules.
}
